The Data Loader

The DataLoader in GeometricMachineLearning is designed to make training convenient.

The data loader can be called with various types of arrays as input, for example a snapshot matrix:

SnapshotMatrix = rand(Float32, 10, 100)

dl = DataLoader(SnapshotMatrix)

DataLoader{Float32, Array{Float32, 3}, Nothing, :RegularData}(Float32[0.7935174; 0.512083; … ; 0.8357026; 0.86754715;;; 0.25151742; 0.08026576; … ; 0.36044437; 0.032096624;;; 0.9151157; 0.35054576; … ; 0.6579215; 0.8846297;;; … ;;; 0.28054446; 0.26132202; … ; 0.04970306; 0.6006115;;; 0.60537976; 0.97311807; … ; 0.8523442; 0.7508296;;; 0.7266169; 0.72154105; … ; 0.69915456; 0.07218069], nothing, 10, 1, 100, nothing, nothing)

or a snapshot tensor:

SnapshotTensor = rand(Float32, 10, 100, 5)

dl = DataLoader(SnapshotTensor)

DataLoader{Float32, Array{Float32, 3}, Nothing, :TimeSeries}(Float32[0.12884498 0.15011185 … 0.7972284 0.11691469; 0.12181258 0.83533436 … 0.21483058 0.7992308; … ; 0.996816 0.55073285 … 0.51858366 0.14735049; 0.84645814 0.7408406 … 0.22798508 0.5168506;;; 0.39642513 0.6985275 … 0.3815114 0.82186097; 0.28746855 0.13175249 … 0.781827 0.15540057; … ; 0.112520576 0.39290076 … 0.35496205 0.14164674; 0.99315476 0.49614573 … 0.7260494 0.935468;;; 0.95325583 0.56795347 … 0.8280192 0.9575122; 0.79654855 0.044080436 … 0.31166393 0.92006266; … ; 0.3793484 0.8703349 … 0.7191165 0.7904116; 0.9216456 0.9240375 … 0.28107548 0.66940105;;; 0.47643054 0.21802682 … 0.8984381 0.50347847; 0.9097196 0.5234011 … 0.6907321 0.018257558; … ; 0.7972778 0.5326306 … 0.69055957 0.70865357; 0.6227588 0.3435256 … 0.07462472 0.103679895;;; 0.64061177 0.41235352 … 0.8691766 0.13222212; 0.6292822 0.26100695 … 0.56183195 0.92388135; … ; 0.20890659 0.7560639 … 0.42843163 0.124678254; 0.8821233 0.77076703 … 0.26841605 0.7945069], nothing, 10, 100, 5, nothing, nothing)

Here the DataLoader has different properties :RegularData and :TimeSeries. This indicates that in the first case we treat all columns in the input tensor independently (this is mostly used for autoencoder problems), whereas in the second case we have time series-like data, which are mostly used for integration problems. The default when using a matrix is :RegularData and the default when using a tensor is :TimeSeries.

We can also treat a problem with a matrix as input as a time series-like problem by providing an additional keyword argument: autoencoder=false:

SnapshotMatrix = rand(Float32, 10, 100)

dl = DataLoader(SnapshotMatrix; autoencoder=false)
dl.input_time_steps

100

If we deal with hamiltonian systems we typically split the coordinates into a $q$ and a $p$ part. Such data can also be used as input arguments for DataLoader:

SymplecticSnapshotTensor = (q = rand(Float32, 10, 100, 5), p = rand(Float32, 10, 100, 5))

dl = DataLoader(SymplecticSnapshotTensor)

DataLoader{Float32, @NamedTuple{q::Array{Float32, 3}, p::Array{Float32, 3}}, Nothing, :TimeSeries}((q = Float32[0.60178727 0.5188358 … 0.5687976 0.5258379; 0.6246846 0.120066226 … 0.48057485 0.06811714; … ; 0.6122685 0.65196365 … 0.89803195 0.594159; 0.11581069 0.00013953447 … 0.14773953 0.9272346;;; 0.6899029 0.34674406 … 0.31619626 0.4410538; 0.5335458 0.2067284 … 0.86850536 0.5993938; … ; 0.7206255 0.7507658 … 0.60065097 0.095531106; 0.05072975 0.61222374 … 0.8029081 0.39673692;;; 0.3355788 0.064177275 … 0.57838047 0.046132326; 0.099368036 0.88524413 … 0.28851467 0.24302214; … ; 0.2739632 0.41952735 … 0.04038292 0.17282218; 0.9945615 0.5308265 … 0.9017268 0.67036057;;; 0.655416 0.8131216 … 0.62286013 0.39744252; 0.27395374 0.93823665 … 0.24637294 0.6105824; … ; 0.6778358 0.76958936 … 0.024114251 0.25363266; 0.05339408 0.40876502 … 0.06097424 0.99030757;;; 0.76529753 0.3001309 … 0.16931349 0.052681983; 0.74563026 0.33851433 … 0.3934309 0.75838494; … ; 0.21679318 0.45867747 … 0.28936237 0.91129; 0.6673674 0.047843397 … 0.33775532 0.13353497], p = Float32[0.035817206 0.8362933 … 0.23499113 0.07134402; 0.68971205 0.2354604 … 0.644054 0.74157745; … ; 0.37022907 0.4656846 … 0.5739302 0.8862032; 0.09594542 0.3735746 … 0.54163784 0.85833913;;; 0.5252169 0.74354374 … 0.42312032 0.17478806; 0.6690274 0.50128824 … 0.91390413 0.061904192; … ; 0.8980464 0.9669421 … 0.0696488 0.06477487; 0.43105602 0.20286292 … 0.3899536 0.5144353;;; 0.38514966 0.7102506 … 0.43808413 0.35781878; 0.89918923 0.79661167 … 0.5595634 0.5250484; … ; 0.9608298 0.904929 … 0.39743143 0.2420097; 0.19046605 0.5761797 … 0.60255325 0.70568746;;; 0.4643318 0.27843875 … 0.10912591 0.18404788; 0.11293244 0.20125198 … 0.02741623 0.65951437; … ; 0.5852866 0.81234527 … 0.5748584 0.7557259; 0.69047844 0.8046138 … 0.55655384 0.27888268;;; 0.8214506 0.8902004 … 0.8684449 0.07575405; 0.13672042 0.4240499 … 0.91588897 0.058788955; … ; 0.65794104 0.7743629 … 0.6321687 0.05691576; 0.69926107 0.64254737 … 0.34735453 0.15000474]), nothing, 20, 100, 5, nothing, nothing)

The dimension of the system is then the sum of the dimensions of the $q$ and the $p$ component:

dl.input_dim

20

The Batch struct

If we want to draw mini batches from a data loader, we need to allocate an instance of Batch. If we call the corresponding functor on an instance of DataLoader we get the following result:

matrix_data = rand(Float32, 2, 10)
dl = DataLoader(matrix_data; autoencoder = true)

batch = Batch(3)
batch(dl)

([(1, 1), (1, 8), (1, 6)], [(1, 7), (1, 9), (1, 2)], [(1, 10), (1, 5), (1, 3)], [(1, 4)])

The output of applying the batch functor is always of the form:

\[([(b_{1,1}^t, b_{1,1}^p), (b_{1,2}^t, b_{1,2}^p), \ldots], [(b_{2,1}^t, b_{2, 1}^p), (b_{2, 2}^t, b_{2, 2}^p), \ldots], [(b_{3, 1}^t, b_{3, 2}^p), \ldots], \ldots),\]

so it is a tuple of vectors of tuples. One vector represents one batch. the tuples in one vector always have two entries: a time index and a parameter index, indicated by the superscripts $t$ and $p$ respectively in the equation above.

This also works if the data are in $(q, p)$ form:

qp_data = (q = rand(Float32, 2, 10), p = rand(Float32, 2, 10))
dl = DataLoader(qp_data; autoencoder = true)

batch = Batch(3)
batch(dl)

([(1, 5), (1, 6), (1, 4)], [(1, 2), (1, 7), (1, 3)], [(1, 10), (1, 8), (1, 9)], [(1, 1)])

In those two examples the autoencoder keyword was set to true (the default). This is why the first index was always 1. This changes if we set autoencoder = false:

qp_data = (q = rand(Float32, 2, 10), p = rand(Float32, 2, 10))
dl = DataLoader(qp_data; autoencoder = false) # false is default

batch = Batch(3)
batch(dl)

([(6, 1), (7, 1), (2, 1)], [(5, 1), (8, 1), (3, 1)], [(1, 1), (4, 1), (9, 1)])

Specifically the routines do the following:

$\mathtt{n\_indices}\leftarrow \mathtt{n\_params}\lor\mathtt{input\_time\_steps},$

$\mathtt{indices} \leftarrow \mathtt{shuffle}(\mathtt{1:\mathtt{n\_indices}}),$

$\mathcal{I}_i \leftarrow \mathtt{indices[(i - 1)} \cdot \mathtt{batch\_size} + 1 \mathtt{:} i \cdot \mathtt{batch\_size]}\text{ for }i=1, \ldots, (\mathrm{last} -1),$

$\mathcal{I}_\mathtt{last} \leftarrow \mathtt{indices[}(\mathtt{n\_batches} - 1) \cdot \mathtt{batch\_size} + 1\mathtt{:end]}.$

Note that the routines are implemented in such a way that no two indices appear double, i.e. we sample without replacement.

Sampling from a tensor

We can also sample tensor data.

qp_data = (q = rand(Float32, 2, 20, 3), p = rand(Float32, 2, 20, 3))
dl = DataLoader(qp_data)

# also specify sequence length here
batch = Batch(4, 5)
batch(dl)

([(8, 2), (3, 2), (5, 2), (6, 2)], [(9, 2), (11, 2), (1, 2), (2, 2)], [(7, 2), (10, 2), (4, 2), (8, 1)], [(3, 1), (5, 1), (6, 1), (9, 1)], [(11, 1), (1, 1), (2, 1), (7, 1)], [(10, 1), (4, 1), (8, 3), (3, 3)], [(5, 3), (6, 3), (9, 3), (11, 3)], [(1, 3), (2, 3), (7, 3), (10, 3)], [(4, 3)])

Sampling from a tensor is done the following way ($\mathcal{I}_i$ again denotes the batch indices for the $i$-th batch):

$\mathtt{time\_indices} \leftarrow \mathtt{shuffle}(\mathtt{1:}(\mathtt{input\_time\_steps} - \mathtt{seq\_length} - \mathtt{prediction_window}),$

$\mathtt{parameter\_indices} \leftarrow \mathtt{shuffle}(\mathtt{1:n\_params}),$

$\mathtt{complete\_indices} \leftarrow \mathtt{product}(\mathtt{time\_indices}, \mathtt{parameter\_indices}),$

$\mathcal{I}_i \leftarrow \mathtt{complete\_indices[}(i - 1) \cdot \mathtt{batch\_size} + 1 : i \cdot \mathtt{batch\_size]}\text{ for }i=1, \ldots, (\mathrm{last} -1),$

$\mathcal{I}_\mathrm{last} \leftarrow \mathtt{complete\_indices[}(\mathrm{last} - 1) \cdot \mathtt{batch\_size} + 1\mathtt{:end]}.$

This algorithm can be visualized the following way (here batch_size = 4):

Here the sampling is performed over the second axis (the time step dimension) and the third axis (the parameter dimension). Whereas each block has thickness 1 in the $x$ direction (i.e. pertains to a single parameter), its length in the $y$ direction is seq_length. In total we sample as many such blocks as the batch size is big. By construction those blocks are never the same throughout a training epoch but may intersect each other!

Library Functions

GeometricMachineLearning.DataLoader — Type

DataLoader(data)

Make an instance based on a data set.

This is designed to make training convenient.

Fields of DataLoader

The fields of the DataLoader struct are the following:

input: The input data with axes (i) system dimension, (ii) number of time steps and (iii) number of parameters.

output: The tensor that contains the output (supervised learning) - this may be of type Nothing if the constructor is only called with one tensor (unsupervised learning).

input_dim: The dimension of the system, i.e. what is taken as input by a regular neural network.

input_time_steps: The length of the entire time series (length of the second axis).

n_params: The number of parameters that are present in the data set (length of third axis)

output_dim: The dimension of the output tensor (first axis). If output is of type Nothing, then this is also of type Nothing.

output_time_steps: The size of the second axis of the output tensor. If output is of type Nothing, then this is also of type Nothing.

Implementation

Even though DataLoader can be called with inputs of various forms, internally it always stores tensors with three axes.

using GeometricMachineLearning

data = [1 2 3; 4 5 6]
dl = DataLoader(data)
dl.input

# output

[ Info: You have provided a matrix as input. The axes will be interpreted as (i) system dimension and (ii) number of parameters.
2×1×3 Array{Int64, 3}:
[:, :, 1] =
 1
 4

[:, :, 2] =
 2
 5

[:, :, 3] =
 3
 6

GeometricMachineLearning.Batch — Type

Batch is a struct whose functor acts on an instance of DataLoader to produce a sequence of training samples for training for one epoch.

See Batch(::Int), Batch(::Int, ::Int) and Batch(::Int, ::Int, ::Int) for the different constructors.

The functor

An instance of Batch can be called on an instance of DataLoader to produce a sequence of samples that contain all the input data, i.e. for training for one epoch.

The output of applying batch:Batch to dl::DataLoader is a tuple of vectors of integers. Each of these vectors contains two integers: the first is the time index and the second one is the parameter index.

Examples

Consider the following example for drawing batches of size 2 for an instance of DataLoader constructed with a vector:

using GeometricMachineLearning
import Random

Random.seed!(123)

dl = DataLoader(rand(5))
batch = Batch(2)

batch(dl)

# output

[ Info: You have provided a matrix as input. The axes will be interpreted as (i) system dimension and (ii) number of parameters.
([(1, 4), (1, 3)], [(1, 2), (1, 1)], [(1, 5)])

Here the first index is always 1 (the time dimension). We get a total number of 3 batches. The last batch is only of size 1 because we sample without replacement. Also see the docstring for DataLoader(::AbstractVector).

GeometricMachineLearning.number_of_batches — Function

number_of_batches(dl, batch)

Compute the number of batches.

Here the big distinction is between data that are time-series like and data that are autoencoder like.

GeometricMachineLearning.convert_input_and_batch_indices_to_array — Function

convert_input_and_batch_indices_to_array(dl, batch, batch_indices)

Assign batch data based on batch indices.

Examples

using GeometricMachineLearning

dl = DataLoader([0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9])
batch = Batch(3)
batch_indices = [(1, 1), (1, 3), (1, 5)]

GeometricMachineLearning.convert_input_and_batch_indices_to_array(dl, batch, batch_indices)

# output

[ Info: You have provided a matrix as input. The axes will be interpreted as (i) system dimension and (ii) number of parameters.
1×1×3 Array{Float64, 3}:
[:, :, 1] =
 0.1

[:, :, 2] =
 0.3

[:, :, 3] =
 0.5