GeometricMachineLearning Library Functions

AbstractNeuralNetworks.Chain — Method

Chain can also be called with a neural network as input.

AbstractNeuralNetworks.Chain — Method

Build a chain for an LASympnet for which init_upper_linear is false and init_upper_act is false.

AbstractNeuralNetworks.Chain — Method

Build a chain for an LASympnet for which init_upper_linear is false and init_upper_act is true.

AbstractNeuralNetworks.Chain — Method

Build a chain for an LASympnet for which init_upper_linear is true and init_upper_act is false.

AbstractNeuralNetworks.Chain — Method

Build a chain for an LASympnet for which init_upper_linear is true and init_upper_act is true.

GeometricMachineLearning.AbstractCache — Type

AbstractCache has subtypes:

AdamCache

MomentumCache

GradientCache

BFGSCache

All of them can be initialized with providing an array (also supporting manifold types).

GeometricMachineLearning.AbstractLieAlgHorMatrix — Type

AbstractLieAlgHorMatrix is a supertype for various horizontal components of Lie algebras. We usually call this $\mathfrak{g}^\mathrm{hor}$.

GeometricMachineLearning.AbstractRetraction — Type

AbstractRetraction is a type that comprises all retraction methods for manifolds. For every manifold layer one has to specify a retraction method that takes the layer and elements of the (global) tangent space.

GeometricMachineLearning.ActivationLayer — Type

ActivationLayer is the struct corresponding to the constructors ActivationLayerQ and ActivationLayerP. See those for more information.

GeometricMachineLearning.ActivationLayerP — Method

Performs:

\[\begin{pmatrix} q \\ p \end{pmatrix} \mapsto \begin{pmatrix} q \\ p + \mathrm{diag}(a)\sigma(q) \end{pmatrix}.\]

GeometricMachineLearning.ActivationLayerQ — Method

Performs:

\[\begin{pmatrix} q \\ p \end{pmatrix} \mapsto \begin{pmatrix} q + \mathrm{diag}(a)\sigma(p) \\ p \end{pmatrix}.\]

GeometricMachineLearning.AdamOptimizer — Type

Defines the Adam Optimizer. Algorithm and suggested defaults are taken from [27] (page 301).

GeometricMachineLearning.AdamOptimizerWithDecay — Type

Defines the Adam Optimizer with weight decay.

Constructors

The default constructor takes as input:

n_epochs::Int

η₁: the learning rate at the start

η₂: the learning rate at the end

ρ₁: the decay parameter for the first moment

ρ₂: the decay parameter for the second moment

δ: the safety parameter

T (keyword argument): the type.

The second constructor is called with:

n_epochs::Int

T

... the rest are keyword arguments

GeometricMachineLearning.AutoEncoder — Type

The autoencoder architecture

An autoencoder [27] is a neural network consisting of an encoder $\Psi^e$ and a decoder $\Psi^d$. In the simplest case they are trained on some data set $\mathcal{D}$ to reduce the following error:

\[||\Psi^d\circ\Psi^e(\mathcal{D}) - \mathcal{D}||,\]

which we call the reconstruction error or autoencoder error (see the docs for AutoEncoderLoss) and $||\cdot||$ is some norm.

Implementation details.

Abstract AutoEncoder type. If a custom <:AutoEncoder architecture is implemented it should have the fields full_dim, reduced_dim, n_encoder_blocks and n_decoder_blocks. Further the routines encoder, decoder, encoder_parameters and decoder_parameters should be extended.

GeometricMachineLearning.AutoEncoderLoss — Type

This loss should always be used together with a neural network of type AutoEncoder (and it is also the default for training such a network).

It simply computes:

\[\mathtt{AutoEncoderLoss}(nn\mathtt{::Loss}, input) = ||nn(input) - input||.\]

GeometricMachineLearning.BFGSCache — Type

The cache for the BFGS optimizer.

It stores an array for the previous time step B and the inverse of the Hessian matrix H.

It is important to note that setting up this cache already requires a derivative! This is not the case for the other optimizers.

GeometricMachineLearning.BFGSDummyCache — Type

In order to initialize BGGSCache we first need gradient information. This is why we initially have this BFGSDummyCache until gradient information is available.

NOTE: we may not need this.

GeometricMachineLearning.BFGSOptimizer — Type

This is an implementation of the Broyden-Fletcher-Goldfarb-Shanno (BFGS) optimizer.

GeometricMachineLearning.Batch — Type

Batch is a struct whose functor acts on an instance of DataLoader to produce a sequence of training samples for training for one epoch.

The Constructor

The constructor for Batch is called with:

batch_size::Int

seq_length::Int (optional)

prediction_window::Int (optional)

The first one of these arguments is required; it indicates the number of training samples in a batch. If we deal with time series data then we can additionaly supply a sequence length and a prediction window as input arguments to Batch. These indicate the number of input vectors and the number of output vectors.

The functor

An instance of Batch can be called on an instance of DataLoader to produce a sequence of samples that contain all the input data, i.e. for training for one epoch. The output of applying batch:Batch to dl::DataLoader is a tuple of vectors of integers. Each of these vectors contains two integers: the first is the time index and the second one is the parameter index.

GeometricMachineLearning.BiasLayer — Type

A bias layer that does nothing more than add a vector to the input. This is needed for LA-SympNets.

GeometricMachineLearning.Classification — Type

Classification Layer that takes a matrix as an input and returns a vector that is used for MNIST classification.

It has the following arguments:

M: input dimension

N: output dimension

activation: the activation function

And the following optional argument:

average: If this is set to true, then the output is computed as $\frac{1}{N}\sum_{i=1}^N[input]_{\bullet{}i}$. If set to false (the default) it picks the last column of the input.

GeometricMachineLearning.ClassificationTransformer — Type

This is a transformer neural network for classification purposes. At the moment this is only used for training on MNIST, but can in theory be used for any classification problem.

It has to be called with a DataLoader that stores an input and an output tensor. The optional arguments are:

n_heads: The number of heads in the MultiHeadAttention (mha) layers. Default: 7.

n_layers: The number of transformer layers. Default: 16.

activation: The activation function. Default: softmax.

Stiefel: Wheter the matrices in the mha layers are on the Stiefel manifold.

add_connection: Whether the input is appended to the output of the mha layer. (skip connection)

GeometricMachineLearning.DataLoader — Type

Data Loader is a struct that creates an instance based on a tensor (or different input format) and is designed to make training convenient.

Constructor

The data loader can be called with various inputs:

A single vector: If the data loader is called with a single vector (and no other arguments are given), then this is interpreted as an autoencoder problem, i.e. the second axis indicates parameter values and/or time steps and the system has a single degree of freedom (i.e. the system dimension is one).

A single matrix: If the data loader is called with a single matrix (and no other arguments are given), then this is interpreted as an autoencoder problem, i.e. the first axis is assumed to indicate the degrees of freedom of the system and the second axis indicates parameter values and/or time steps.

A single tensor: If the data loader is called with a single tensor, then this is interpreted as an integration problem with the second axis indicating the time step and the third one indicating the parameters.

A tensor and a vector: This is a special case (MNIST classification problem). For the MNIST problem for example the input are $n_p$ matrices (first input argument) and $n_p$ integers (second input argument).

A NamedTuple with fields q and p: The NamedTuple contains (i) two matrices or (ii) two tensors.

An EnsembleSolution: The EnsembleSolution typically comes from GeometricProblems.

When we supply a single vector or a single matrix as input to DataLoader and further set autoencoder = false (keyword argument), then the data are stored as an integration problem and the second axis is assumed to indicate time steps.

Fields of DataLoader

The fields of the DataLoader struct are the following:

input: The input data with axes (i) system dimension, (ii) number of time steps and (iii) number of parameters.

output: The tensor that contains the output (supervised learning) - this may be of type Nothing if the constructor is only called with one tensor (unsupervised learning).

input_dim: The dimension of the system, i.e. what is taken as input by a regular neural network.

input_time_steps: The length of the entire time series (length of the second axis).

n_params: The number of parameters that are present in the data set (length of third axis)

output_dim: The dimension of the output tensor (first axis). If output is of type Nothing, then this is also of type Nothing.

output_time_steps: The size of the second axis of the output tensor. If output is of type Nothing, then this is also of type Nothing.

The input and output fields of DataLoader

Even though the arguments to the Constructor may be vectors or matrices, internally DataLoader always stores tensors.

GeometricMachineLearning.DataLoader — Method

Data Loader is a struct that creates an instance based on a tensor (or different input format) and is designed to make training convenient.

Constructor

The data loader can be called with various inputs:

A single vector: If the data loader is called with a single vector (and no other arguments are given), then this is interpreted as an autoencoder problem, i.e. the second axis indicates parameter values and/or time steps and the system has a single degree of freedom (i.e. the system dimension is one).

A single matrix: If the data loader is called with a single matrix (and no other arguments are given), then this is interpreted as an autoencoder problem, i.e. the first axis is assumed to indicate the degrees of freedom of the system and the second axis indicates parameter values and/or time steps.

A single tensor: If the data loader is called with a single tensor, then this is interpreted as an integration problem with the second axis indicating the time step and the third one indicating the parameters.

A tensor and a vector: This is a special case (MNIST classification problem). For the MNIST problem for example the input are $n_p$ matrices (first input argument) and $n_p$ integers (second input argument).

A NamedTuple with fields q and p: The NamedTuple contains (i) two matrices or (ii) two tensors.

An EnsembleSolution: The EnsembleSolution typically comes from GeometricProblems.

When we supply a single vector or a single matrix as input to DataLoader and further set autoencoder = false (keyword argument), then the data are stored as an integration problem and the second axis is assumed to indicate time steps.

GeometricMachineLearning.DataLoader — Method

Constructor for EnsembleSolution from package GeometricSolutions with field q.

GeometricMachineLearning.DataLoader — Method

Constructor for EnsembleSolution form package GeometricSolutions with fields q and p.

GeometricMachineLearning.Decoder — Type

Abstract Decoder type. If a custom <:Decoder architecture is implemented it should have the fields full_dim, reduced_dim and n_decoder_blocks.

GeometricMachineLearning.Encoder — Type

Abstract Encoder type. If a custom <:Encoder architecture is implemented it should have the fields full_dim, reduced_dim and n_encoder_blocks.

GeometricMachineLearning.GSympNet — Type

GSympNet is called with a single input argument, the system dimension, or with an instance of DataLoader. Optional input arguments are:

upscaling_dimension::Int: The upscaling dimension of the gradient layer. See the documentation for GradientLayerQ and GradientLayerP for further explanation. The default is 2*dim.

nhidden::Int: The number of hidden layers (i.e. layers that are not input or output layers). The default is 2.

activation: The activation function that is applied. By default this is tanh.

init_upper::Bool: Initialize the gradient layer so that it first modifies the $q$-component. The default is true.

GeometricMachineLearning.GlobalSection — Type

This implements global sections for the Stiefel manifold and the Symplectic Stiefel manifold.

In practice this is implemented using Householder reflections, with the auxiliary column vectors given by: |0| |0| |.| |1| ith spot for i in (n+1) to N (or with random columns) |0| |.| |0|

Maybe consider dividing the output in the check functions by n!

Implement a general global section here!!!! Tₓ𝔐 → G×𝔤 !!!!!! (think about random initialization!)

GeometricMachineLearning.GradientLayer — Type

GradientLayer is the struct corresponding to the constructors GradientLayerQ and GradientLayerP. See those for more information.

GeometricMachineLearning.GradientLayerP — Method

The gradient layer that changes the $q$ component. It is of the form:

\[\begin{bmatrix} \mathbb{I} & \mathbb{O} \\ \nabla{}V & \mathbb{I} \end{bmatrix},\]

with $V(p) = \sum_{i=1}^Ma_i\Sigma(\sum_jk_{ij}p_j+b_i)$, where $\Sigma$ is the antiderivative of the activation function $\sigma$ (one-layer neural network). We refer to $M$ as the upscaling dimension. Such layers are by construction symplectic.

GeometricMachineLearning.GradientLayerQ — Method

The gradient layer that changes the $q$ component. It is of the form:

\[\begin{bmatrix} \mathbb{I} & \nabla{}V \\ \mathbb{O} & \mathbb{I} \end{bmatrix},\]

with $V(p) = \sum_{i=1}^Ma_i\Sigma(\sum_jk_{ij}p_j+b_i)$, where $\Sigma$ is the antiderivative of the activation function $\sigma$ (one-layer neural network). We refer to $M$ as the upscaling dimension. Such layers are by construction symplectic.

GeometricMachineLearning.GradientOptimizer — Type

Define the Gradient optimizer, i.e. W ← W - η*∇f(W) Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.GrassmannLayer — Type

Defines a layer that performs simple multiplication with an element of the Grassmann manifold.

GeometricMachineLearning.GrassmannLieAlgHorMatrix — Type

This implements the horizontal component of a Lie algebra that is isomorphic to the Grassmann manifold.

GeometricMachineLearning.GrassmannManifold — Type

The GrassmannManifold is based on the StiefelManifold

GeometricMachineLearning.HRedSys — Type

HRedSys computes the reconstructed dynamics in the full system based on the reduced one. Optionally it can be compared to the FOM solution.

It can be called using the following constructor: HRedSys(N, n; encoder, decoder, v_full, f_full, v_reduced, f_reduced, parameters, tspan, tstep, ics, projection_error) where

encoder: a function $\mathbb{R}^{2N}\mapsto{}\mathbb{R}^{2n}$

decoder: a (differentiable) function $\mathbb{R}^{2n}\mapsto\mathbb{R}^{2N}$

v_full: a (differentiable) mapping defined the same way as in GeometricIntegrators.

f_full: a (differentiable) mapping defined the same way as in GeometricIntegrators.

v_reduced: a (differentiable) mapping defined the same way as in GeometricIntegrators.

f_reduced: a (differentiable) mapping defined the same way as in GeometricIntegrators.

parameters: a NamedTuple that parametrizes the vector fields (the same for fullvectorfield and reducedvectorfield)

tspan: a tuple (t₀, tₗ) that specifies start and end point of the time interval over which integration is performed.

tstep: the time step

ics: the initial condition for the big system.

projection_error: the error $||M - \mathcal{R}\circ\mathcal{P}(M)||$ where $M$ is the snapshot matrix; $\mathcal{P}$ and $\mathcal{R}$ are the reduction and reconstruction respectively.

GeometricMachineLearning.LASympNet — Type

LASympNet is called with a single input argument, the system dimension, or with an instance of DataLoader. Optional input arguments are:

depth::Int: The number of linear layers that are applied. The default is 5.

nhidden::Int: The number of hidden layers (i.e. layers that are not input or output layers). The default is 2.

activation: The activation function that is applied. By default this is tanh.

init_upper_linear::Bool: Initialize the linear layer so that it first modifies the $q$-component. The default is true.

init_upper_act::Bool: Initialize the activation layer so that it first modifies the $q$-component. The default is true.

GeometricMachineLearning.LayerWithManifold — Type

Additional types to make handling manifolds more readable.

GeometricMachineLearning.LinearLayer — Type

LinearLayer is the struct corresponding to the constructors LinearLayerQ and LinearLayerP. See those for more information.

GeometricMachineLearning.LinearLayerP — Method

Equivalent to a left multiplication by the matrix:

\[\begin{pmatrix} \mathbb{I} & \mathbb{O} \\ B & \mathbb{I} \end{pmatrix}, \]

where $B$ is a symmetric matrix.

GeometricMachineLearning.LinearLayerQ — Method

Equivalent to a left multiplication by the matrix:

\[\begin{pmatrix} \mathbb{I} & B \\ \mathbb{O} & \mathbb{I} \end{pmatrix}, \]

where $B$ is a symmetric matrix.

GeometricMachineLearning.LowerTriangular — Type

A lower-triangular matrix is an $n\times{}n$ matrix that has ones on the diagonal and zeros on the upper triangular.

The data are stored in a vector $S$ similarly to SkewSymMatrix.

The struct two fields: S and n. The first stores all the entries of the matrix in a sparse fashion (in a vector) and the second is the dimension $n$ for $A\in\mathbb{R}^{n\times{}n}$.

GeometricMachineLearning.Manifold — Type

rand is implemented for manifolds that use the initialization of the StiefelManifold and the GrassmannManifold by default.

GeometricMachineLearning.ManifoldLayer — Type

This defines a manifold layer that only has one matrix-valued manifold $A$ associated with it does $x\mapsto{}Ax$.

GeometricMachineLearning.MomentumOptimizer — Type

Define the Momentum optimizer, i.e. V ← αV - ∇f(W) W ← W + ηV Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.MultiHeadAttention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

Constructor

Takes input arguments:

dim::Int: The system dimension

n_heads::Int: The number of heads.

Stiefel::Bool=true (keyword argument): whether the weights should be put on the Stiefel manifold.

retraction::AbstractRetraction (keyword argument): what kind of retraction should be used. By default this is the geodesic retraction.

add_connection::Bool=true (keyword argument): determines if the input should be added to the output for the final result.

GeometricMachineLearning.Optimizer — Type

Optimizer struct that stores the 'method' (i.e. Adam with corresponding hyperparameters), the cache and the optimization step.

It takes as input an optimization method and the parameters of a network.

For technical reasons we first specify an OptimizerMethod that stores all the hyperparameters of the optimizer.

GeometricMachineLearning.Optimizer — Method

A functor for Optimizer. It is called with: - nn::NeuralNetwork - dl::DataLoader - batch::Batch - n_epochs::Int - loss

The last argument is a function through which Zygote differentiates. This argument is optional; if it is not supplied GeometricMachineLearning defaults to an appropriate loss for the DataLoader.

GeometricMachineLearning.Optimizer — Method

Typically the Optimizer is not initialized with the network parameters, but instead with a NeuralNetwork struct.

GeometricMachineLearning.PSDArch — Type

The architecture

Proper symplectic decomposition (PSD) can be seen as a SymplecticAutoencoder for which the decoder and the encoder are both PSD-like matrices (see the docs for PSDLayer.

Training

For optimizing the parameters in this architecture no neural network training is necessary (see the docs for solve!).

The constructor

The constructor only takes two arguments as input:

full_dim::Integer

reduced_dim::Integer

GeometricMachineLearning.PSDLayer — Type

This is a PSD-like layer used for symplectic autoencoders. One layer has the following shape:

\[A = \begin{bmatrix} \Phi & \mathbb{O} \\ \mathbb{O} & \Phi \end{bmatrix},\]

where $\Phi$ is an element of the Stiefel manifold $St(n, N)$.

The constructor of PSDLayer is called by PSDLayer(M, N; retraction=retraction):

M is the input dimension.

N is the output dimension.

retraction is an instance of a struct with supertype AbstractRetraction. The only options at the moment are Geodesic() and Cayley().

GeometricMachineLearning.RegularTransformerIntegrator — Type

The regular transformer used as an integrator (multi-step method).

The constructor is called with the following arguments:

sys_dim::Int

transformer_dim::Int: the default is transformer_dim = sys_dim.

n_blocks::Int: The default is 1.

n_heads::Int: the number of heads in the multihead attentio layer (default is n_heads = sys_dim)

L::Int the number of transformer blocks (default is L = 2).

upscaling_activation: by default identity

resnet_activation: by default tanh

add_connection:Bool=true (keyword argument): if the input should be added to the output.

GeometricMachineLearning.SkewSymMatrix — Type

A SkewSymMatrix is a matrix $A$ s.t. $A^T = -A$.

If the constructor is called with a matrix as input it returns a symmetric matrix via the projection $A \mapsto \frac{1}{2}(A - A^T)$. This is a projection defined via the canonical metric $\mathbb{R}^{n\times{}n}\times\mathbb{R}^{n\times{}n}\to\mathbb{R}, (A,B) \mapsto \mathrm{Tr}(A^TB)$.

The first index is the row index, the second one the column index.

The struct two fields: S and n. The first stores all the entries of the matrix in a sparse fashion (in a vector) and the second is the dimension $n$ for $A\in\mathbb{R}^{n\times{}n}$.

GeometricMachineLearning.StiefelLayer — Type

Defines a layer that performs simple multiplication with an element of the Stiefel manifold.

GeometricMachineLearning.StiefelLieAlgHorMatrix — Type

StiefelLieAlgHorMatrix is the horizontal component of the Lie algebra of skew-symmetric matrices (with respect to the canonical metric). The projection here is: (\pi:S \to SE ) where

\[E = \begin{pmatrix} \mathbb{I}_{n} \\ \mathbb{O}_{(N-n)\times{}n} \end{pmatrix}.\]

The matrix (E) is implemented under StiefelProjection in GeometricMachineLearning.

An element of StiefelLieAlgMatrix takes the form:

\[\begin{pmatrix} A & B^T \\ B & \mathbb{O} \end{pmatrix},\]

where (A) is skew-symmetric (this is SkewSymMatrix in GeometricMachineLearning).

If the constructor is called with a big (N\times{}N) matrix, then the projection is performed the following way:

\[\begin{pmatrix} A & B_1 \\ B_2 & D \end{pmatrix} \mapsto \begin{pmatrix} \mathrm{skew}(A) & -B_2^T \\ B_2 & \mathbb{O} \end{pmatrix}.\]

The operation $\mathrm{skew}:\mathbb{R}^{n\times{}n}\to\mathcal{S}_\mathrm{skew}(n)$ is the skew-symmetrization operation. This is equivalent to calling the constructor of SkewSymMatrix with an (n\times{}n) matrix.

GeometricMachineLearning.StiefelManifold — Type

An implementation of the Stiefel manifold. It has various convenience functions associated with it:

check

rand

rgrad

metric

GeometricMachineLearning.StiefelProjection — Type

Outer constructor for StiefelProjection. This works with two integers as input and optionally the type.

GeometricMachineLearning.StiefelProjection — Type

An array that essentially does vcat(I(n), zeros(N-n, n)) with GPU support. It has three inner constructors. The first one is called with the following arguments:

backend: backends as supported by KernelAbstractions.

T::Type

N::Integer

n::Integer

The second constructor is called by supplying a matrix as input. The constructor will then extract the backend, the type and the dimensions of that matrix.

The third constructor is called by supplying an instance of StiefelLieAlgHorMatrix.

Technically this should be a subtype of StiefelManifold.

GeometricMachineLearning.SymmetricMatrix — Type

A SymmetricMatrix $A$ is a matrix $A^T = A$.

If the constructor is called with a matrix as input it returns a symmetric matrix via the projection:

\[A \mapsto \frac{1}{2}(A + A^T).\]

This is a projection defined via the canonical metric $(A,B) \mapsto \mathrm{tr}(A^TB)$.

Internally the struct saves a vector $S$ of size $n(n+1)\div2$. The conversion is done the following way:

\[[A]_{ij} = \begin{cases} S[( (i-1) i ) \div 2 + j] & \text{if $i\geq{}j$}\\ S[( (j-1) j ) \div 2 + i] & \text{else}. \end{cases}\]

So $S$ stores a string of vectors taken from $A$: $S = [\tilde{a}_1, \tilde{a}_2, \ldots, \tilde{a}_n]$ with $\tilde{a}_i = [[A]_{i1},[A]_{i2},\ldots,[A]_{ii}]$.

GeometricMachineLearning.SympNet — Type

SympNet type encompasses GSympNets and LASympnets.

TODO: -[ ] add bias to LASympNet!

GeometricMachineLearning.SympNetLayer — Type

Implements the various layers from the SympNet paper: (https://www.sciencedirect.com/science/article/abs/pii/S0893608020303063). This is a super type of Gradient, Activation and Linear.

For the linear layer, the activation and the bias are left out, and for the activation layer $K$ and $b$ are left out!

GeometricMachineLearning.SympNetLayer — Method

This is called when a SympnetLayer is applied to a NamedTuple. It calls apply_layer_to_nt_and_return_array.

GeometricMachineLearning.SymplecticAutoencoder — Type

The architecture

The symplectic autoencoder architecture was introduced in [28]. Like any other autoencoder it consists of an encoder $\Psi^e:\mathbb{R}^{2N}\to\mathbb{R}^{2n}$ and a decoder $\Psi^d:\mathbb{R}^{2n}\to\mathbb{R}^{2N}$ with $n\ll{}N$. These satisfy the following properties:

\[\nabla_z\Psi^e\mathbb{J}_{2N}(\nabla_z\Psi^e\mathbb{J}_{2N})^T = \mathbb{J}_{2n} \text{ and } (\nabla_\xi\Psi^d)^T\mathbb{J}_{2N}\nabla_\xi\Psi^d = \mathbb{J}_{2n}.\]

Because the decoder has this particular property, the reduced system can be described by the Hamiltonian $H\circ\Psi^d$:

\[\mathbb{J}_{2n}\nabla_\xi(H\circ\Psi^d) = \mathbb{J}_{2n}(\nabla_\xi\Psi^d)^T\nabla_{\Psi^d(\xi)}H = \mathbb{J}_{2n}(\nabla_\xi\Psi^d)^T\mathbb{J}_{2N}^T\mathbb{J}_{2N}\nabla_{\Psi^d(\xi)}H = (\nabla_\xi\Psi^d)^+X_H(\Psi^d(\xi)),\]

where $(\nabla_\xi\Psi^d)^+$ is the pseudoinverse of $\nabla_\xi\Psi^d$ (for more details see the docs on the AutoEncoder type).

The constructor

The constructor is called with

full_dim::Integer

reduced_dim::Integer

n_encoder_layers::Integer = 4 (keyword argument)

n_encoder_blocks::Integer = 2 (keyword argument)

n_decoder_layers::Integer = 1 (keyword argument)

n_decoder_blocks::Integer = 3 (keyword argument)

sympnet_upscale::Integer = 5 (keyword argument)

activation = tanh (keyword argument)

encoder_init_q::Bool = true (keyword argument)

decoder_init_q::Bool = true (keyword argument)

GeometricMachineLearning.SymplecticPotential — Type

SymplecticPotential(n)

Returns a symplectic matrix of size 2n x 2n

\[\begin{pmatrix} \mathbb{O} & \mathbb{I} \\ \mathbb{O} & -\mathbb{I} \\ \end{pmatrix}\]

GeometricMachineLearning.TrainingData — Type

TrainingData stores:

 - problem 

 - shape 

 - get 

 - symbols 

 - dim 

 - noisemaker

GeometricMachineLearning.TransformerIntegrator — Type

Encompasses various transformer architectures, such as the structure-preserving transformer and the linear symplectic transformer.

GeometricMachineLearning.TransformerLoss — Type

The loss for a transformer network (especially a transformer integrator). The constructor is called with:

seq_length::Int

prediction_window::Int (default is 1).

GeometricMachineLearning.UpperTriangular — Type

An upper-triangular matrix is an $n\times{}n$ matrix that has ones on the diagonal and zeros on the upper triangular.

The data are stored in a vector $S$ similarly to SkewSymMatrix.

The struct two fields: S and n. The first stores all the entries of the matrix in a sparse fashion (in a vector) and the second is the dimension $n$ for $A\in\mathbb{R}^{n\times{}n}$.

GeometricMachineLearning.VolumePreservingAttention — Type

Volume-preserving attention (single head attention)

Drawbacks:

the super fast activation is only implemented for sequence lengths of 2, 3, 4 and 5.

other sequence lengths only work on CPU for now (lu decomposition has to be implemented to work for tensors in parallel).

Constructor

The constructor is called with:

dim::Int: The system dimension

seq_length::Int: The sequence length to be considered. The default is zero, i.e. arbitrary sequence lengths; this works for all sequence lengths but doesn't apply the super-fast activation.

skew_sym::Bool (keyword argument): specifies if we the weight matrix is skew symmetric or arbitrary (default is false).

Functor

Applying a layer of type VolumePreservingAttention does the following:

First we perform the operation $X \mapsto X^T A X =: C$, where $X\in\mathbb{R}^{N\times\mathtt{seq\_length}}$ is a vector containing time series data and $A$ is the skew symmetric matrix associated with the layer.

In a second step we compute the Cayley transform of $C$; $\Lambda = \mathrm{Cayley}(C)$.

The output of the layer is then $X\Lambda$.

GeometricMachineLearning.VolumePreservingFeedForward — Type

Realizes a volume-preserving neural network as a combination of VolumePreservingLowerLayer and VolumePreservingUpperLayer.

Constructor

The constructor is called with the following arguments:

sys_dim::Int: The system dimension.

n_blocks::Int: The number of blocks in the neural network (containing linear layers and nonlinear layers). Default is 1.

n_linear::Int: The number of linear VolumePreservingLowerLayers and VolumePreservingUpperLayers in one block. Default is 1.

activation: The activation function for the nonlinear layers in a block.

init_upper::Bool=false (keyword argument): Specifies if the first layer is lower or upper.

GeometricMachineLearning.VolumePreservingFeedForwardLayer — Type

Super-type of VolumePreservingLowerLayer and VolumePreservingUpperLayer. The layers do the following:

\[x \mapsto \begin{cases} \sigma(Lx + b) & \text{where $L$ is }\mathtt{LowerTriangular} \\ \sigma(Ux + b) & \text{where $U$ is }\mathtt{UpperTriangular}. \end{cases}\]

The functor can be applied to a vecotr, a matrix or a tensor.

Constructor

The constructors are called with:

sys_dim::Int: the system dimension.

activation=tanh: the activation function.

include_bias::Bool=true (keyword argument): specifies whether a bias should be used.

GeometricMachineLearning.VolumePreservingLowerLayer — Type

See the documentation for VolumePreservingFeedForwardLayer.

GeometricMachineLearning.VolumePreservingTransformer — Type

The volume-preserving transformer with the Cayley activation function and built-in upscaling. The arguments for the constructor are:

sys_dim::Int

seq_length::Int: The sequence length of the data fed into the transformer.

n_blocks::Int=1 (keyword argument): The number of blocks in one transformer unit (containing linear layers and nonlinear layers). Default is 1.

n_linear::Int=1 (keyword argument): The number of linear VolumePreservingLowerLayers and VolumePreservingUpperLayers in one block. Default is 1.

L::Int=1 (keyword argument): The number of transformer units (default is 2).

activation=tanh (keyward argument): The activation function (tanh by default).

init_upper::Bool=false (keyword argument): Specifies if the network first acts on the $q$ component.

skew_sym::Bool=false (keyword argument): specifies if we the weight matrix is skew symmetric or arbitrary.

GeometricMachineLearning.VolumePreservingUpperLayer — Type

See the documentation for VolumePreservingFeedForwardLayer.

AbstractNeuralNetworks.update! — Method

Optimization for an entire neural networks with BFGS. What is different in this case is that we still have to initialize the cache.

If o.step == 1, then we initialize the cache

Base.iterate — Method

This function computes a trajectory for a Transformer that has already been trained for valuation purposes.

It takes as input:

nn: a NeuralNetwork (that has been trained).

ics: initial conditions (a matrix in $\mathbb{R}^{2n\times\mathtt{seq\_length}}$ or NamedTuple of two matrices in $\mathbb{R}^{n\times\mathtt{seq\_length}}$)

n_points::Int=100 (keyword argument): The number of steps for which we run the prediction.

prediction_window::Int=size(ics.q, 2): The prediction window (i.e. the number of steps we predict into the future) is equal to the sequence length (i.e. the number of input time steps) by default.

Base.iterate — Method

This function computes a trajectory for a SympNet that has already been trained for valuation purposes.

It takes as input:

nn: a NeuralNetwork (that has been trained).

ics: initial conditions (a NamedTuple of two vectors)

Base.vec — Method

If vec is applied onto Triangular, then the output is the associated vector.

Base.vec — Method

If vec is applied onto SkewSymMatrix, then the output is the associated vector.

GeometricMachineLearning.Gradient — Function

This is an old constructor and will be depricated. For change_q=true it is equivalent to GradientLayerQ; for change_q=false it is equivalent to GradientLayerP.

If full_grad=false then ActivationLayer is called

GeometricMachineLearning.Transformer — Method

The architecture for a "transformer encoder" is essentially taken from arXiv:2010.11929, but with the difference that no layer normalization is employed. This is because we still need to find a generalization of layer normalization to manifolds.

The transformer is called with the following inputs:

dim: the dimension of the transformer

n_heads: the number of heads

L: the number of transformer blocks

In addition we have the following optional arguments:

activation: the activation function used for the ResNet (tanh by default)

Stiefel::Bool: if the matrices $P^V$, $P^Q$ and $P^K$ should live on a manifold (false by default)

retraction: which retraction should be used (Geodesic() by default)

add_connection::Bool: if the input should by added to the ouput after the MultiHeadAttention layer is used (true by default)

use_bias::Bool: If the ResNet should use a bias (true by default)

GeometricMachineLearning.accuracy — Method

Computes the accuracy (as opposed to the loss) of a neural network classifier.

It takes as input:

model::Chain

ps: parameters of the network

dl::DataLoader

GeometricMachineLearning.apply_layer_to_nt_and_return_array — Method

This function is used in the wrappers where the input to the SympNet layers is not a NamedTuple (as it should be) but an AbstractArray.

It converts the Array to a NamedTuple (via assign_q_and_p), then calls the SympNet routine(s) and converts back to an AbstractArray (with vcat).

GeometricMachineLearning.assign_batch_kernel! — Method

Takes as input a batch tensor (to which the data are assigned), the whole data tensor and two vectors params and time_steps that include the specific parameters and time steps we want to assign.

Note that this assigns sequential data! For e.g. being processed by a transformer.

GeometricMachineLearning.assign_output_estimate — Method

The function assign_output_estimate is closely related to the transformer. It takes the last prediction_window columns of the output and uses them for the final prediction. i.e.

\[\mathbb{R}^{N\times\mathtt{pw}}\to\mathbb{R}^{N\times\mathtt{pw}}, \begin{bmatrix} z^{(1)}_1 & \cdots & z^{(T)}_1 \\ \cdots & \cdots & \cdots \\ z^{(1)}_n & \cdots & z^{(T})_n \end{bmatrix} \mapsto \begin{bmatrix} z^{(T - \mathtt{pw})}_1 & \cdots & z^{(T)}_1 \\ \cdots & \cdots & \cdots \\ z^{(T - \mathtt{pw})}_n & \cdots & z^{(T})_n\end{bmatrix} \]

GeometricMachineLearning.assign_output_kernel! — Method

This should be used together with assign_batch_kernel!. It assigns the corresponding output (i.e. target).

GeometricMachineLearning.assign_q_and_p — Method

Allocates two new arrays q and p whose first dimension is half of that of the input x. This should also be supplied through the second argument N.

The output is a Tuple containing q and p.

GeometricMachineLearning.augment_zeros_kernel! — Method

Used for differentiating assignoutputestimate (this appears in the loss).

GeometricMachineLearning.build_v_reduced — Method

Builds the reduced vector field based on the full vector field for a Hamiltonian system. We derive the reduced vector field via the reduced Hamiltonian: $\tilde{H} := H\circ\Psi^\mathrm{dec}$. We then get

\[\mathbb{J}_{2n}\nabla_\xi\tilde{H} = \mathbb{J}_{2n}(\nabla\Psi^\mathrm{dec})^T\mathbb{J}_{2N}^T\mathbb{J}_{2N}\nabla_z{}H = \mathbb{J}_{2n}(\nabla\Psi^\mathrm{dec})^T\mathbb{J}_{2N}^T \begin{pmatrix} v(z) \\ f(z) \end{pmatrix} = \begin{pmatrix} - (\nabla_p\Psi_q)^Tf(z) + (\nabla_p\Psi_p)^Tv(z) \\ (\nabla_q\Psi_q)^Tf(z) - (\nabla_q\Psi_p)^Tv(z) \end{pmatrix}.\]

GeometricMachineLearning.compute_iterations — Method

This function gives iterations from the full dimension to the reduced dimension (i.e. the intermediate steps). The iterations are given in ascending order.

GeometricMachineLearning.compute_iterations_for_symplectic_system — Method

This function gives iterations from the full dimension to the reduced dimension (i.e. the intermediate steps). The iterations are given in ascending order. Only even steps are allowed here.

GeometricMachineLearning.compute_output_of_mha — Method

Applies MHA to an abstract matrix. This is the same independent of whether the input is added to the output or not.

GeometricMachineLearning.convert_input_and_batch_indices_to_array — Method

Takes the output of the batch functor and uses it to create the corresponding array (NamedTuples).

GeometricMachineLearning.convert_input_and_batch_indices_to_array — Method

Takes the output of the batch functor and uses it to create the corresponding array.

GeometricMachineLearning.crop_array_for_transformer_loss — Method

This crops the output array of the neural network so that it conforms with the output it should be compared to. This is needed for the transformer loss.

GeometricMachineLearning.custom_mat_mul — Method

Multiplies a matrix with a vector, a matrix or a tensor.

GeometricMachineLearning.decoder_layers_from_iteration — Method

Takes as input the autoencoder architecture and a vector of integers specifying the layer dimensions in the decoder. Has to return a tuple of AbstractExplicitLayers.

GeometricMachineLearning.draw_batch! — Method

This assigns the batch if the data are in form of a matrix.

GeometricMachineLearning.encoder_layers_from_iteration — Method

Takes as input the autoencoder architecture and a vector of integers specifying the layer dimensions in the encoder. Has to return a tuple of AbstractExplicitLayers.

GeometricMachineLearning.init_optimizer_cache — Method

Wrapper for the functions setup_adam_cache, setup_momentum_cache, setup_gradient_cache, setup_bfgs_cache. These appear outside of optimizer_caches.jl because the OptimizerMethods first have to be defined.

GeometricMachineLearning.initialize_hessian_inverse — Method

This initializes the inverse of the Hessian for various arrays. This requires an implementation of a vectorization operation vec. This is important for custom arrays.

GeometricMachineLearning.map_index_for_symplectic_potential — Method

This assigns the right index for the symplectic potential. To be used with assign_ones_for_symplectic_potential_kernel!.

GeometricMachineLearning.mat_tensor_mul — Method

Extend mat_tensor_mul to a multiplication by the adjoint of an element of StiefelManifold.

GeometricMachineLearning.mat_tensor_mul — Method

Extend mat_tensor_mul to a multiplication by an element of StiefelManifold.

GeometricMachineLearning.metric — Method

Implements the canonical Riemannian metric for the Stiefel manifold:

\[g_Y: (\Delta_1, \Delta_2) \mapsto \mathrm{tr}(\Delta_1^T(\mathbb{I} - \frac{1}{2}YY^T)\Delta_2).\]

It is called with:

Y::StiefelManifold

Δ₁::AbstractMatrix

Δ₂::AbstractMatrix`

GeometricMachineLearning.number_of_batches — Method

Gives the number of batches. Inputs are of type DataLoader and Batch.

Here the big distinction is between data that are time-series like and data that are autoencoder like.

GeometricMachineLearning.onehotbatch — Method

One-hot-batch encoding of a vector of integers: $input\in\{0,1,\ldots,9\}^\ell$. The output is a tensor of shape $10\times1\times\ell$.

\[0 \mapsto \begin{bmatrix} 1 & 0 & \ldots & 0 \end{bmatrix}.\]

In more abstract terms: $i \mapsto e_i$.

GeometricMachineLearning.optimization_step! — Method

Optimization for an entire neural network, the way this function should be called.

inputs:

o::Optimizer

model::Chain

ps::Tuple

dx::Tuple

GeometricMachineLearning.optimization_step! — Method

Optimization for a single layer.

inputs:

o::Optimizer

d::Union{AbstractExplicitLayer, AbstractExplicitCell}

ps::NamedTuple: the parameters

C::NamedTuple: NamedTuple of the caches

dx::NamedTuple: NamedTuple of the derivatives (output of AD routine)

ps, C and dx must have the same keys.

GeometricMachineLearning.optimize_for_one_epoch! — Method

Optimize for an entire epoch. For this you have to supply:

an instance of the optimizer.

the neural network model

the parameters of the model

the data (in form of DataLoader)

in instance of Batch that contains batch_size (and optionally seq_length)

With the optional argument:

the loss, which takes the model, the parameters ps and an instance of DataLoader as input.

The output of optimize_for_one_epoch! is the average loss over all batches of the epoch:

\[output = \frac{1}{\mathtt{steps\_per\_epoch}}\sum_{t=1}^\mathtt{steps\_per\_epoch}loss(\theta^{(t-1)}).\]

This is done because any reverse differentiation routine always has two outputs: a pullback and the value of the function it is differentiating. In the case of zygote: loss_value, pullback = Zygote.pullback(ps -> loss(ps), ps) (if the loss only depends on the parameters).

GeometricMachineLearning.patch_index — Method

Based on coordinates i,j this returns the batch index (for MNIST data set for now).

GeometricMachineLearning.rgrad — Method

Computes the Riemannian gradient for the Stiefel manifold given an element $Y\in{}St(N,n)$ and a matrix $\nabla{}L\in\mathbb{R}^{N\times{}n}$ (the Euclidean gradient). It computes the Riemannian gradient with respect to the canonical metric (see the documentation for the function metric for an explanation of this). The precise form of the mapping is:

\[\mathtt{rgrad}(Y, \nabla{}L) \mapsto \nabla{}L - Y(\nabla{}L)^TY\]

It is called with inputs:

Y::StiefelManifold

e_grad::AbstractMatrix: i.e. the Euclidean gradient (what was called $\nabla{}L$) above.

GeometricMachineLearning.solve! — Method

PSDArch does not require neural network training since it is a strictly linear operation that can be solved with singular value decomposition (SVD).

GeometricMachineLearning.split_and_flatten — Method

split_and_flatten takes a tensor as input and produces another one as output (essentially rearranges the input data in an intricate way) so that it can easily be processed with a transformer.

The optional arguments are:

patch_length: by default this is 7.

number_of_patches: by default this is 16.

GeometricMachineLearning.tensor_mat_skew_sym_assign — Method

Takes as input:

Z::AbstractArray{T, 3}: A tensor that stores a bunch of time series.

A::AbstractMatrix: A matrix that is used to perform various scalar products.

For one of these time series the function performs the following computation:

\[ (z^{(i)}, z^{(j)}) \mapsto (z^{(i)})^TAz^{(j)} \text{ for } i > j.\]

The result of this are $n(n-2)\div2$ scalar products. These scalar products are written into a lower-triangular matrix and the final output of the function is a tensor of these lower-triangular matrices.

GeometricMachineLearning.tensor_mat_skew_sym_assign_kernel! — Method

A kernel that computes the weighted scalar products of all combinations of vectors in the matrix Z except where the two vectors are the same and writes the result into a tensor of skew symmetric matrices C.

GeometricMachineLearning.train! — Function

train!(...)

Perform a training of a neural networks on data using given method a training Method

Different ways of use:

train!(neuralnetwork, data, optimizer = GradientOptimizer(1e-2), training_method; nruns = 1000, batch_size = default(data, type), showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data : the data (see TrainingData)

optimizer = GradientOptimizer: the optimization method (see Optimizer)

training_method : specify the loss function used

nruns : number of iteration through the process with default value

batch_size : size of batch of data used for each step

GeometricMachineLearning.train! — Method

train!(neuralnetwork, data, optimizer, training_method; nruns = 1000, batch_size, showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data::AbstractTrainingData : the data

``

GeometricMachineLearning.within_patch_index — Method

Based on coordinates i,j this returns the index within the batch

GeometricMachineLearning.write_ones_kernel! — Method

Kernel that is needed for functions relating to SymmetricMatrix and SkewSymMatrix