(Matrix) Manifolds

Manifolds are topological spaces that locally look like vector spaces. In the following we restrict ourselves to finite-dimensional smooth[1] manifolds. In this section we routinely denote points on a manifold by lower case letters like $x, y$ and $z$ if we speak about general properties and by upper case letters like $A$ and $B$ if we talk about specific examples of matrix manifolds. All manifolds that can be used to build neural networks in GeometricMachineLearning, such as the Stiefel manifold and the Grassmann manifold are matrix manifolds.

Definition

A finite-dimensional smooth manifold of dimension $n$ is a second-countable Hausdorff space $\mathcal{M}$ for which $\forall{}x\in\mathcal{M}$ we can find a neighborhood $U$ that contains $x$ and a corresponding homeomorphism $\varphi_U:U\cong{}W\subset\mathbb{R}^n$ where $W$ is an open subset. The homeomorphisms $\varphi_U$ are referred to as coordinate charts. If two such coordinate charts overlap, i.e. if $U_1\cap{}U_2\neq\{\}$, then the map $\varphi_{U_2}^{-1}\circ\varphi_{U_1}$ has to be $C^\infty$.

One example of a manifold that is also important for GeometricMachineLearning is the Lie group[2] of orthonormal matrices $SO(N)$. Before we can proof that $SO(N)$ is a manifold we first need the preimage theorem.

The Preimage Theorem

Before we can state the preimage theorem we need another definition[3]:

Definition

Consider a smooth mapping $g: \mathcal{M}\to\mathcal{N}$ from one manifold to another. A point $y\in\mathcal{N}$ is called a regular value of $g$ if $\forall{}x\in{}g^{-1}\{y\}$ the map $T_xg:T_A\mathcal{M}\to{}T_{y}\mathcal{N}$ is surjective.

We now state the preimage theorem:

Theorem (Preimage Theorem)

Consider a smooth map $g:\mathcal{M}\to\mathcal{N}$ from one manifold to another (we assume the dimensions of the two manifolds to be $m+n$ and $m$ respectively). Then the preimage of a regular point $y$ of $\mathcal{N}$ is a submanifold of $\mathcal{M}$. Furthermore the codimension of $g^{-1}\{y\}$ is equal to the dimension of $\mathcal{N}$ and the tangent space $T_x(g^{-1}\{y\})$ is equal to the kernel of $T_xg$.

Proof

Because $\mathcal{N}$ has manifold structure we can find a chart $\varphi_U:U\to\mathbb{R}^m$ for some neighborhood $U$ that contains $y$. We further consider a point $A\in{}g^{-1}\{y\}$ and a chart around it $\psi_V:V\to\mathbb{R}^{m+n}$. By the implicit function theorem we can then find a mapping $h$ that turns $\varphi_U\circ{}g\circ\psi_V^{-1}$ into a projection $(x_1, \ldots, x_{n+m}) \mapsto (x_{n+1}, \ldots, x_{n+m})$. We now consider the neighborhood $V_1\times\{0\} = \psi(V \cup f^{-1}\{y\})$ for $\psi(V) = V_1\times{}V_2$ with the coordinate chart $(x_1, \ldots, x_n) \mapsto \psi(x_1, \ldots, x_n, 0, \ldots, 0).$ As this map is also smooth by the implicit function theorem this proofs our assertion.

Example

The group $SO(N)$ is a Lie group (i.e. has manifold structure).

Proof

The vector space $\mathbb{R}^{N\times{}N}$ clearly has manifold structure. The group $SO(N)$ is equivalent to one of the level sets of the mapping: $g:\mathbb{R}^{N\times{}N}\to\mathcal{S}(N), A\mapsto{}A^TA - \mathbb{I}$, i.e. it is the component of $f^{-1}\{\mathbb{I}\}$ that contains $\mathbb{I}$. We still need to proof that $\mathbb{I}$ is a regular point of $g$, i.e. that for $A\in{}SO(N)$ the mapping $T_Ag$ is surjective. This means that $\forall{}B\in\mathcal{S}(N), A\in\mathbb{R}^{N\times{}N}$ $\exists{}C\in\mathbb{R}^{N\times{}N}$ s.t. $C^TA + A^TC = B$. The element $C=\frac{1}{2}AB\in\mathcal{R}^{N\times{}N}$ satisfies this property.

Similarly we can also proof:

Example

The sphere $S^n:=\{x\in\mathbb{R}^{n+1}: x^Tx = 1\}$ is a manifold of dimension $n$.

Proof

Take $g(x) = x^x - 1$ and proceed as in the case of $SO(N)$.

Note that both these manifolds, $SO(N)$ and $S^n$ are matrix manifolds, i.e. an element of $\mathcal{M}$ can be written as an element of $\mathbb{R}^{N\times{}N}$ in the first case and $\mathbb{R}^{n\times{}1}$ in the second case. The additional conditions we impose on these manifolds are $A^TA = \mathbb{I}$ in the first case and $x^Tx = 1$ in the second case. Both of these manifolds belong to the category of Stiefel manifolds.

Tangent Spaces

A tangent space can be seen as the collection of all possible velocities a curve can take at a point on a manifold. For this consider a manifold $\mathcal{M}$ and a point $x$ on it and the collection of $C^\infty$ curves through $x$:

Definition

A mapping $\gamma:(-\epsilon, \epsilon)\to\mathcal{M}$ that is $C^\infty$ and for which we have $\gamma(0) = x$ is called a $C^\infty$ curve through $x$.

The tangent space of $\mathcal{M}$ at $x$ is the collection of the first derivatives of all $\gamma$:

Definition

The tangent space of $\mathcal{M}$ at $x$ is the collection of all $C^\infty$ curves at $x$ modulo the equivalence class $\gamma_1 \sim \gamma_2 \iff \gamma_1'(0) = \gamma_2'(0)$. It is denoted by $T_x\mathcal{M}$.

As is customary we write $[\gamma]$ for the equivalence class of $\gamma$ and this is by definition equivalent to $\gamma'(0)$. The tangent space $T_x\mathcal{M}$ can be shown to be homeomorphic[4] to $\mathbb{R}^n$ where $n$ is the dimension of the manifold $\mathcal{M}$. If the homeomorphism is constructed through the coordinate chart $(\varphi, U)$ we call it $\varphi'(x)$ or simply[5] $\varphi'$. If we are given a map $g:\mathcal{M}\to\mathcal{N}$ we further define $T_xg = (\varphi')^{-1}\circ(\varphi\circ{}g\psi^{-1})'\circ{}\psi'$, i.e. a smooth map between two manifolds $\mathcal{M}$ and $\mathcal{N}$ induces a smooth map between the tangent spaces $T_x\mathcal{M}$ and $T_{g(x)}\mathcal{N}$.

We want to demonstrate this principle of constructing the tangent space from curves through the example of $S^2$. We consider the following curves:

$\gamma_1(t) = \begin{pmatrix} 0 \\ sin(t) \\ cos(t) \end{pmatrix},$

$\gamma_2(t) = \begin{pmatrix} sin(t) \\ 0 \\ cos(t) \end{pmatrix},$

$\gamma_3(t) = \begin{pmatrix} \exp(-t ^ 2 / 2) t \sin(t) \\ \exp(-t ^ 2 / 2) t cos(t) \\ \sqrt{1 - (t ^ 2) exp(-t^2)} \end{pmatrix}.$

We now plot the manifold $S^2$, the three curves described above and the associated tangent vectors (visualized as arrows). Note that the tangent vectors induced by $\gamma_1$ and $\gamma_3$ are the same; for these curves we have $\gamma_1 \sim \gamma_3$ and the tangent vectors of those two curves coincide:

The tangent space $T_x\mathcal{M}$ for

\[x = \begin{pmatrix}0 \\ 0 \\ 1 \end{pmatrix} \]

is also shown.

Vector Fields

A time-independent vector field[6] is an object that specifies a velocity for every point on a domain. We first give the definition of a vector field on the vector space $\mathbb{R}^n$ and limit ourselves here to $C^\infty$ vector fields:

Definition

A vector field on $\mathbb{R}^n$ is a smooth map $X:\mathbb{R}^n\to\mathbb{R}^n$.

The definition of a vector field on a manifold is not much more complicated:

Definition

A vector field on $\mathcal{M}$ is a map $X$ defined on $\mathcal{M}$ such that $X(x)\in{}T_x\mathcal{M}$ and $\varphi'\circ{}X\circ(\varphi)^{-1}$ is smooth for any coordinate chart $(\varphi, U)$ that contains $x$.

In the section on the existence-and-uniqueness theorem we show that every vector field has a unique solution given an initial condition; i.e. given a point $x\in\mathcal{M}$ and a vector field $X$ we can find a curve $\gamma$ such that $\gamma(0) = x$ and $\gamma'(t) = X(\gamma(t))$ for all $t$ in some interval $(-\epsilon, \epsilon)$.

The Tangent Bundle

To each manifold $\mathcal{M}$ we can associate another manifold which we call the tangent bundle and denote by $T\mathcal{M}$. The points on this manifold are:

\[T\mathcal{M} = \{ (x, v_x): x\in\mathcal{M},\, v_x\in{}T_x\mathcal{M} \}.\]

Coordinate charts on this manifold can be constructed in a straightforward manner; for every coordinate chart $\varphi_U$ the map $\varphi_U'(x)$ gives a homeomorphism between $T_x\mathcal{M}$ and $\mathbb{R}^n$for any $x\in{}U$. We can then find a neighborhood of any point $(x, v_x)$ by taking $\pi^{-1}(U) = \{(x, v_x): x\in{}U, v_x\in{}T_x\mathcal{M}\}$ and this neighborhood is isomorphic to $\mathbb{R}^{2n}$ via $(x, v_x) \mapsto (\varphi_U(x), \varphi'(x)v_x)$. The geodesic spray is an important vector field defined on $T\mathcal{M}$.

Library Functions

GeometricMachineLearning.Manifold — Type

A manifold in GeometricMachineLearning is a sutype of AbstractMatrix. All manifolds are matrix manifolds and therefore stored as matrices. More details can be found in the docstrings for the StiefelManifold and the GrassmannManifold.

References

[10]

P.-A. Absil, R. Mahony and R. Sepulchre. Optimization algorithms on matrix manifolds (Princeton University Press, Princeton, New Jersey, 2008).

1Smooth here means $C^\infty$.

2Lie groups are manifolds that also have a group structure, i.e. there is an operation $\mathcal{M}\times\mathcal{M}\to\mathcal{M},(a,b)\mapsto{}ab$ s.t. $(ab)c = a(bc)$ and there exists a neutral element$e\mathcal{M}$ s.t. $ae$ = $a$ $\forall{}a\in\mathcal{M}$ as well as an (for every $a$) inverse element $a^{-1}$ s.t. $a(a^{-1}) = e$. The neutral element $e$ we refer to as $\mathbb{I}$ when dealing with matrix manifolds.

3In this definition we use the notation $T_xg$. This will be explained below. For we will interpret $T_xg$ simply as $(\varphi_U\circ{}g\circ\psi_V^{-1})'$ where $\varphi_U$ is a coordinate chart around $y = g(x)$ and $\psi_V$ is a coordinate chart around $x$.

4Note that we have not formally defined addition for $T_x\mathcal{M}$. This can be done through the definition $[\gamma] + [\beta] = [\alpha]$ where $\alpha$ is any $C^\infty$ curve through $x$ that satisfies $\alpha'(0) = \beta(0) + \gamma(0)$. Note that we can always find such an $\alpha$ by the existence and uniqueness theorem.

5We will further discuss this when we introduce the tangent bundle.

6Also called ordinary differential equation (ODE).