Hamiltonian Model Order Reduction

Hamiltonian PDEs are partial differential equations that, like its ODE counterpart, have a Hamiltonian associated with it. The linear wave equation can be written as such a Hamiltonian PDE with

\[\mathcal{H}(q, p; \mu) := \frac{1}{2}\int_\Omega\mu^2(\partial_\xi{}q(t,\xi;\mu))^2 + p(t,\xi;\mu)^2d\xi.\]

Note that in contrast to the ODE case where the Hamiltonian is a function $H(\cdot; \mu):\mathbb{R}^{2d}\to\mathbb{R},$ we now have a functional $\mathcal{H}(\cdot, \cdot; \mu):\mathcal{C}^\infty(\mathcal{D})\times\mathcal{C}^\infty(\mathcal{D})\to\mathbb{R}.$ The PDE for to this Hamiltonian can be obtained similarly as in the ODE case:

\[\partial_t{}q(t,\xi;\mu) = \frac{\delta{}\mathcal{H}}{\delta{}p} = p(t,\xi;\mu), \quad \partial_t{}p(t,\xi;\mu) = -\frac{\delta{}\mathcal{H}}{\delta{}q} = \mu^2\partial_{\xi{}\xi}q(t,\xi;\mu)\]

Neglecting the Hamiltonian structure of a system can have grave consequences on the performance of the reduced order model [46–48] which is why all algorithms in GeometricMachineLearning designed for producing reduced order models respect the structure of the system.

The Symplectic Solution Manifold

As with regular parametric PDEs, we also associate a solution manifold with Hamiltonian PDEs. This is a finite-dimensional manifold, on which the dynamics can be described through a Hamiltonian ODE. The reduced system, with which we approximate this symplectic solution manifold, is a low dimensional symplectic vector space $\mathbb{R}^{2n}$ together with a reduction $\mathcal{P}$ and a reconstruction $\mathcal{R}.$ If we now take an initial condition on the solution manifold $\hat{u}_0\in\mathcal{M} \approx \mathcal{R}(\mathbb{R}^{2n})$ and project it to the reduced space with $\mathcal{P}$, we get $u = \mathcal{P}(\hat{u}_0).$ We can now integrate it on the reduced space via the induced differential equation, which is of canonical Hamiltonian form, and obtain an orbit $u(t)$ which can then be mapped back to an orbit on the solution manifold[4] via $\mathcal{R}.$ The resulting orbit $\mathcal{R}(u(t))$ is ideally the unique orbit on the full order model $\hat{u}(t)\in\mathcal{M}$.

For Hamiltonian model order reduction we additionally require that the reduction $\mathcal{P}$ satisfies

\[ \nabla_z\mathcal{P}\mathbb{J}_{2N}(\nabla_z\mathcal{P})^T = \mathbb{J}_{2N} \text{ for $z\in\mathbb{R}^{2N}$}\]

and the reconstruction $\mathcal{R}$ satisfies[5]

\[ (\nabla_z\mathcal{R})^T\mathbb{J}_{2N}\nabla_z\mathcal{R} = \mathbb{J}_{2n}.\]

With this we have

Theorem

A Hamiltonian system on the reduced space $(\mathbb{R}^{2n}, \mathbb{J}_{2n}^T)$ is equivalent to a non-canonical symplectic system on $(\mathcal{M}, \mathbb{J}_{2N}^T|_\mathcal{M})$ where

\[ \mathcal{M} = \mathcal{R}(\mathbb{R}^{2n})\]

is an approximation to the solution manifold.

For the proof we use the fact that $\mathcal{M} = \mathcal{R}(\mathbb{R}^{2n})$ is a manifold whose coordinate chart is the local inverse of $\mathcal{R}$ which we will call $\psi$, i.e. around a point $y\in\mathcal{M}$ we have $\mathcal{R}\circ\psi(y) = y$.

Proof

Note that the tangent space at $y = \mathcal{R}(z)$ to $\mathcal{M}$ is:

\[ T_y\mathcal{M} = \{(\nabla_z\mathcal{R})v: v\in\mathbb{R}^{2n}\}.\]

The mapping

\[ \mathcal{M} \to T\mathcal{M}, y \mapsto (\nabla_z\mathcal{R})\mathbb{J}_{2n}\nabla_zH\]

is clearly a vector field. We now prove that it is symplectic and equal to $\mathbb{J}_{2N}\nabla_y(H\circ\psi).$ For this first note that we have $\mathbb{I} = (\nabla_z\mathcal{R})^+\nabla_z\mathcal{R} = (\nabla_{\mathcal{R}(z)}\psi)\nabla_z\mathcal{R}$ and that the pseudoinverse is unique. We then have:

\[ \mathbb{J}_{2N}\nabla_yH\circ\psi = \mathbb{J}_{2N}(\nabla_y\psi)^T\nabla_{\psi(y)}H = \mathbb{J}_{2N}\left((\nabla_{\psi(y)}\mathcal{R})^+\right)^T\nabla_{\psi(y)}H = (\nabla_{\psi(y)}\mathcal{R})\mathbb{J}_{2n}\nabla_{\psi(y)}H,\]

which proves that every Hamiltonian system on $\mathbb{R}^{2n}$ induces a Hamiltonian system on $\mathcal{M}$. Conversely we built the manifold $\mathcal{M}$ based on the flow of a high-dimensional Hamiltonian system. We can thus approximate the high-dimensional Hamiltonian system with a low-dimensional Hamiltonian system on $\mathbb{R}^{2n}$.

This theorem serves as the basis for Hamiltonian model order reduction via proper symplectic decomposition and symplectic autoencoders.

Proper Symplectic Decomposition

For proper symplectic decomposition (PSD) the reduction $\mathcal{P}$ and the reconstruction $\mathcal{R}$ are constrained to be linear, orthonormal and symplectic. Note that these first two properties are shared with POD. The easiest way[1] to enforce this is through the so-called "cotangent lift" [46]:

\[\mathcal{R} \equiv \Psi_\mathrm{CL} = \begin{bmatrix} \Phi & \mathbb{O} \\ \mathbb{O} & \Phi \end{bmatrix} \text{ where $\Psi_\mathrm{CL}\in{}St(n,N)\subset\mathbb{R}^{N\times{}n}$},\]

i.e. is an element of the Stiefel manifold. If the snapshot matrix is of the form:

\[M = \left[\begin{array}{c:c:c:c} \hat{q}_1(t_0) & \hat{q}_1(t_1) & \quad\ldots\quad & \hat{q}_1(t_f) \\ \hat{q}_2(t_0) & \hat{q}_2(t_1) & \ldots & \hat{q}_2(t_f) \\ \ldots & \ldots & \ldots & \ldots \\ \hat{q}_N(t_0) & \hat{q}_N(t_1) & \ldots & \hat{q}_N(t_f) \\ \hat{p}_1(t_0) & \hat{p}_1(t_1) & \ldots & \hat{p}_1(t_f) \\ \hat{p}_2(t_0) & \hat{p}_2(t_1) & \ldots & \hat{p}_2(t_f) \\ \ldots & \ldots & \ldots & \ldots \\ \hat{p}_{N}(t_0) & \hat{p}_{N}(t_1) & \ldots & \hat{p}_{N}(t_f) \\ \end{array}\right],\]

then $\Phi_\mathrm{CL}$ can be computed in a very straight-forward manner:

Theorem

The ideal cotangent lift $\Psi_\mathrm{CL}$ for the snapshot matrix of form

\[ M = \begin{bmatrix} M_q \\ M_p \end{bmatrix},\]

i.e. the cotangent lift that minimizes the projection error, can be obtained the following way:

Rearrange the rows of the matrix $M$ such that we end up with a $N\times(2\mathtt{nts})$ matrix: $\hat{M} := [M_q, M_p]$.

Perform SVD: $\hat{M} = V\Sigma{}U^T.$

Set $\Phi\gets{}U\mathtt{[:,1:n]}.$

$\Psi_\mathrm{CL}$ is then built based on this $\Phi$.

For details on the cotangent lift (and other methods for linear symplectic model reduction) consult [46]. In GeometricMachineLearning we use the function solve! for this task.

Symplectic Autoencoders

Symplectic Autoencoders are a type of neural network suitable for treating Hamiltonian parametrized PDEs with slowly decaying Kolmogorov $n$-width. It is based on PSD and symplectic neural networks (SympNets).

PSD suffers from the similar shortcomings as regular POD: it is a linear map and the approximation space $\tilde{\mathcal{M}}= \{\Psi^\mathrm{dec}(z_r)\in\mathbb{R}^{2N}:z_r\in\mathrm{R}^{2n}\}$ is therefore also linear. For problems with slowly-decaying Kolmogorov $n$-width this leads to very poor approximations.

In order to overcome this difficulty we use neural networks, more specifically SympNets, together with cotangent lift-like matrices. The resulting architecture, symplectic autoencoders, are discussed in the dedicated section on neural network architectures.

Workflow for Symplectic ROM

As with any other reduced order modeling technique we first discretize the PDE. This should be done with a structure-preserving scheme, thus yielding a (high-dimensional) Hamiltonian ODE as a result. Going back to the example of the linear wave equation, we can discretize this equation with finite differences to obtain a Hamiltonian ODE:

\[\mathcal{H}_\mathrm{discr}(z(t;\mu);\mu) := \frac{1}{2}x(t;\mu)^T\begin{bmatrix} -\mu^2D_{\xi{}\xi} & \mathbb{O} \\ \mathbb{O} & \mathbb{I} \end{bmatrix} x(t;\mu).\]

In Hamiltonian reduced order modelling we try to find a symplectic submanifold in the solution space[2] that captures the dynamics of the full system as well as possible.

Similar to the regular PDE case we again build an encoder $\mathcal{P} \equiv \Psi^\mathrm{enc}$ and a decoder $\mathcal{R} \equiv \Psi^\mathrm{dec}$; but now both these mappings are required to be symplectic!

Concretely this means:

The encoder is a mapping from a high-dimensional symplectic space to a low-dimensional symplectic space, i.e. $\Psi^\mathrm{enc}:\mathbb{R}^{2N}\to\mathbb{R}^{2n}$ such that $\nabla\Psi^\mathrm{enc}\mathbb{J}_{2N}(\nabla\Psi^\mathrm{enc})^T = \mathbb{J}_{2n}$.

The decoder is a mapping from a low-dimensional symplectic space to a high-dimensional symplectic space, i.e. $\Psi^\mathrm{dec}:\mathbb{R}^{2n}\to\mathbb{R}^{2N}$ such that $(\nabla\Psi^\mathrm{dec})^T\mathbb{J}_{2N}\nabla\Psi^\mathrm{dec} = \mathbb{J}_{2n}$.

If these two maps are constrained to linear maps this amounts to PSD.

Library Functions

GeometricMachineLearning.SymplecticEncoder — Type

Abstract SymplecticEncoder type.

See Encoder for the super type and NonLinearSymplecticEncoder for a derived struct.

GeometricMachineLearning.SymplecticDecoder — Type

Abstract SymplecticDecoder type.

See Decoder for the super type and NonLinearSymplecticDecoder for a derived struct.

GeometricMachineLearning.NonLinearSymplecticEncoder — Type

The NonLinearSymplecticEncoder.

This is typically generated by calling encoder on an instance of SymplecticAutoencoder.

GeometricMachineLearning.NonLinearSymplecticDecoder — Type

The NonLinearSymplecticDecoder.

This is typically generated by calling decoder on an instance of SymplecticAutoencoder.

GeometricMachineLearning.HRedSys — Type

HRedSys computes the reconstructed dynamics in the full system based on the reduced one. Optionally it can be compared to the FOM solution.

It can be called using the following constructor: HRedSys(N, n; encoder, decoder, v_full, f_full, v_reduced, f_reduced, parameters, tspan, tstep, ics, projection_error) where

encoder: a function $\mathbb{R}^{2N}\mapsto{}\mathbb{R}^{2n}$

decoder: a (differentiable) function $\mathbb{R}^{2n}\mapsto\mathbb{R}^{2N}$

v_full: a (differentiable) mapping defined the same way as in GeometricIntegrators.

f_full: a (differentiable) mapping defined the same way as in GeometricIntegrators.

v_reduced: a (differentiable) mapping defined the same way as in GeometricIntegrators.

f_reduced: a (differentiable) mapping defined the same way as in GeometricIntegrators.

parameters: a NamedTuple that parametrizes the vector fields (the same for fullvectorfield and reducedvectorfield)

tspan: a tuple (t₀, tₗ) that specifies start and end point of the time interval over which integration is performed.

tstep: the time step

ics: the initial condition for the big system.

projection_error: the error $||M - \mathcal{R}\circ\mathcal{P}(M)||$ where $M$ is the snapshot matrix; $\mathcal{P}$ and $\mathcal{R}$ are the reduction and reconstruction respectively.

GeometricMachineLearning.build_v_reduced — Function

build_v_reduced(v_full, f_full, decoder)

Builds the reduced vector field ($q$ part) based on the full vector field for a Hamiltonian system.

We derive the reduced vector field via the reduced Hamiltonian: $\tilde{H} := H\circ\Psi^\mathrm{dec}$.

We then get

\[\mathbb{J}_{2n}\nabla_\xi\tilde{H} = \mathbb{J}_{2n}(\nabla\Psi^\mathrm{dec})^T\mathbb{J}_{2N}^T\mathbb{J}_{2N}\nabla_z{}H = \mathbb{J}_{2n}(\nabla\Psi^\mathrm{dec})^T\mathbb{J}_{2N}^T \begin{pmatrix} v(z) \\ f(z) \end{pmatrix} = \begin{pmatrix} - (\nabla_p\Psi_q)^Tf(z) + (\nabla_p\Psi_p)^Tv(z) \\ (\nabla_q\Psi_q)^Tf(z) - (\nabla_q\Psi_p)^Tv(z) \end{pmatrix}.\]

build_v_reduced outputs the first half of the entries of this vector field.

GeometricMachineLearning.build_f_reduced — Function

build_f_reduced(v_full, f_full, decoder)

Builds the reduced vector field ($p$ part) based on the full vector field for a Hamiltonian system.

build_f_reduced outputs the second half of the entries of this vector field.

See build_v_reduced for more information.

GeometricMachineLearning.PSDLayer — Type

This is a PSD-like layer used for symplectic autoencoders. One layer has the following shape:

\[A = \begin{bmatrix} \Phi & \mathbb{O} \\ \mathbb{O} & \Phi \end{bmatrix},\]

where $\Phi$ is an element of the Stiefel manifold $St(n, N)$.

The constructor of PSDLayer is called by PSDLayer(M, N; retraction=retraction):

M is the input dimension.

N is the output dimension.

retraction is an instance of a struct with supertype AbstractRetraction. The only options at the moment are Geodesic() and Cayley().

GeometricMachineLearning.PSDArch — Type

The architecture

Proper symplectic decomposition (PSD) can be seen as a SymplecticAutoencoder for which the decoder and the encoder are both PSD-like matrices (see the docs for PSDLayer.

Training

For optimizing the parameters in this architecture no neural network training is necessary (see the docs for solve!).

The constructor

The constructor only takes two arguments as input:

full_dim::Integer

reduced_dim::Integer

GeometricMachineLearning.solve! — Function

PSDArch does not require neural network training since it is a strictly linear operation that can be solved with singular value decomposition (SVD).

References

[48]

P. Buchfink, S. Glas and B. Haasdonk. Symplectic model reduction of Hamiltonian systems on nonlinear manifolds and approximation with weakly symplectic autoencoder. SIAM Journal on Scientific Computing 45, A289–A311 (2023).

[46]

L. Peng and K. Mohseni. Symplectic model reduction of Hamiltonian systems. SIAM Journal on Scientific Computing 38, A1–A27 (2016).

4To be precise, an approximation of the solution manifold $\mathcal{R}(\mathbb{R}^{2n})$, because we are not able to find the solution manifold exactly in practice.

5We should note that satisfying this symplecticity condition is much more important for the reconstruction than for the reduction. We are not entirely sure if the condition on the reduction is really necessary. In [48] it is ignored.

1The original PSD paper [46] proposes another approach to build linear reductions and reconstructions with the so-called "complex SVD." In practice this only brings minor advantages over the cotangent lift however [47].

2The submanifold, that approximates the solution manifold, is $\tilde{\mathcal{M}} = \{\Psi^\mathrm{dec}(z_r)\in\mathbb{R}^{2N}:u_r\in\mathrm{R}^{2n}\}$ where $z_r$ is the reduced state of the system. By a slight abuse of notation we also denote $\mathcal{M}$ by $\mathcal{M}$.