The Data Loader

The DataLoader in GeometricMachineLearning is designed to make training convenient.

The data loader can be called with various types of arrays as input, for example a snapshot matrix:

SnapshotMatrix = [ 1; 2;; 3; 4;; 5; 6;; 7; 8;; 9; 10 ]

dl = DataLoader(SnapshotMatrix; suppress_info = true)

DataLoader{Int64, Array{Int64, 3}, Nothing, :RegularData}([1; 2;;; 3; 4;;; 5; 6;;; 7; 8;;; 9; 10], nothing, 2, 1, 5, nothing, nothing)

or a snapshot tensor:

SnapshotTensor = [ 1;  2;; 3; 4;; 5; 6;; 7; 8;; 9; 10 ;;;]

dl = DataLoader(SnapshotTensor; suppress_info = true)

DataLoader{Int64, Array{Int64, 3}, Nothing, :TimeSeries}([1 3 … 7 9; 2 4 … 8 10;;;], nothing, 2, 5, 1, nothing, nothing)

Here the DataLoader has different properties :RegularData and :TimeSeries. This indicates that in the first case we treat all columns in the input tensor independently; this is mostly used for autoencoder problems. In the second case we have time series-like data, which are mostly used for integration problems. As shown above the default when using a matrix is :RegularData and the default when using a tensor is :TimeSeries.

We can also treat a problem with a matrix as input as a time series-like problem by providing an additional keyword argument: autoencoder=false:

dl = DataLoader(SnapshotMatrix; autoencoder=false, suppress_info = true)
dl |> typeof

DataLoader{Int64, Array{Int64, 3}, Nothing, :TimeSeries}

If we deal with hamiltonian systems we typically split the coordinates into a $q$ and a $p$ part. Such data can also be used as input arguments for DataLoader:

SymplecticSnapshotTensor = (q = SnapshotTensor, p = SnapshotTensor)
dl = DataLoader(SymplecticSnapshotTensor)
dl |> typeof

DataLoader{Int64, @NamedTuple{q::Array{Int64, 3}, p::Array{Int64, 3}}, Nothing, :TimeSeries}

The dimension of the system is then the sum of the dimensions of the $q$ and the $p$ component:

dl.input_dim

4

Drawing Batches with GeometricMachineLearning

If we want to draw mini batches from a data set, we need to allocate an instance of Batch. If we call the corresponding functor on an instance of DataLoader we get the following result[1]:

matrix_data = [ 1 2 3 4  5;
                6 7 8 9 10]
dl = DataLoader(matrix_data; autoencoder = true)

batch = Batch(3)
batches = batch(dl)

([(1, 5), (1, 4), (1, 2)], [(1, 3), (1, 1)])

The output of applying the batch functor is always of the form:

\[([(b_{1,1}^t, b_{1,1}^p), (b_{1,2}^t, b_{1,2}^p), \ldots], [(b_{2,1}^t, b_{2, 1}^p), (b_{2, 2}^t, b_{2, 2}^p), \ldots], [(b_{3, 1}^t, b_{3, 2}^p), \ldots], \ldots),\]

so it is a tuple of vectors of tuples. One vector represents one batch:

for (minibatch, i) in zip(batches[1], axes(batches[1], 1))
    println(stdout, minibatch[1], " = bᵗ₁" * join('₀' + d for d in digits(i)))
    println(stdout, minibatch[2], " = bᵖ₁" * join('₀' + d for d in digits(i)))
    println()
end

1 = bᵗ₁₁
5 = bᵖ₁₁

1 = bᵗ₁₂
4 = bᵖ₁₂

1 = bᵗ₁₃
2 = bᵖ₁₃

The tuples that make up this vector always have two entries: a time index $b^t_{1i}$ and a parameter index $b^p_{1i}$ indicated by the superscripts $t$ and $p$ respectively. Because dl in this example is of autoencoder type, the time index is always one. The parameter index differs. Because the input to DataLoader was a $2\times5$ matrix and we specified a batch size of three, there are two batches in total. The second batch is:

for (minibatch, i) in zip(batches[2], axes(batches[2], 1))
    println(stdout, minibatch[1], " = bᵗ₁" * join('₀' + d for d in digits(i)))
    println(stdout, minibatch[2], " = bᵖ₁" * join('₀' + d for d in digits(i)))
    println()
end

1 = bᵗ₁₁
3 = bᵖ₁₁

1 = bᵗ₁₂
1 = bᵖ₁₂

Looking at the first and the second batch together, we see that we sample with replacement, i.e. all indices $b^p_{1i} = 1, 2, 3, 4, 5$ appear. This also works if the data are in $(q, p)$ form:

qp_data = (q = rand(Float32, 2, 5), p = rand(Float32, 2, 5))
dl = DataLoader(qp_data; autoencoder = true)

batch = Batch(3)
batch(dl)

([(1, 3), (1, 4), (1, 1)], [(1, 2), (1, 5)])

In those two examples the autoencoder keyword was set to true (the default). This is why the first index was always 1. This changes if we set autoencoder = false:

qp_data = (q = rand(Float32, 2, 5), p = rand(Float32, 2, 5))
dl = DataLoader(qp_data; autoencoder = false) # false is default

batch = Batch(3)
batch(dl)

([(4, 1), (3, 1), (2, 1)], [(1, 1)])

Specifically the sampling routines do the following:

$\mathtt{n\_indices}\leftarrow \mathtt{n\_params}\lor\mathtt{input\_time\_steps},$

$\mathtt{indices} \leftarrow \mathtt{shuffle}(\mathtt{1:\mathtt{n\_indices}}),$

$\mathcal{I}_i \leftarrow \mathtt{indices[(i - 1)} \cdot \mathtt{batch\_size} + 1 \mathtt{:} i \cdot \mathtt{batch\_size]}\text{ for }i=1, \ldots, (\mathrm{last} -1),$

$\mathcal{I}_\mathtt{last} \leftarrow \mathtt{indices[}(\mathtt{n\_batches} - 1) \cdot \mathtt{batch\_size} + 1\mathtt{:end]}.$

Note that the routines are implemented in such a way that no two indices appear double, i.e. we sample without replacement.

Sampling from a tensor

We can also sample from a tensor:

qp_data = (q = rand(Float32, 2, 5, 3), p = rand(Float32, 2, 5, 3))
dl = DataLoader(qp_data)

# also specify sequence length and a prediction window here
batch = Batch(4, 2, 0)
batch(dl)

([(1, 3), (3, 3), (2, 3), (4, 3)], [(1, 1), (3, 1), (2, 1), (4, 1)], [(1, 2), (3, 2), (2, 2), (4, 2)])

Sampling from a tensor is done the following way ($\mathcal{I}_i$ again denotes the batch indices for the $i$-th batch):

$\mathtt{time\_indices} \leftarrow \mathtt{shuffle}(\mathtt{1:}(\mathtt{input\_time\_steps} - \mathtt{seq\_length} - \mathtt{prediction\_window}),$

$\mathtt{parameter\_indices} \leftarrow \mathtt{shuffle}(\mathtt{1:n\_params}),$

$\mathtt{complete\_indices} \leftarrow \mathtt{product}(\mathtt{time\_indices}, \mathtt{parameter\_indices}),$

$\mathcal{I}_i \leftarrow \mathtt{complete\_indices[}(i - 1) \cdot \mathtt{batch\_size} + 1 : i \cdot \mathtt{batch\_size]}\text{ for }i=1, \ldots, (\mathrm{last} -1),$

$\mathcal{I}_\mathrm{last} \leftarrow \mathtt{complete\_indices[}(\mathrm{last} - 1) \cdot \mathtt{batch\_size} + 1\mathtt{:end]}.$

Note that we supplied two additional arguments to the Batch constructor here: seq_length and prediction_window. These two arguments specify how many time are considered in one mini batch and how long the prediction runs into the future respectively. These two numbers are explained when we talk about structure on product spaces.

This algorithm can be visualized the following way (here batch_size = 4):

Here the sampling is performed over the second axis (the time step dimension) and the third axis (the parameter dimension). Whereas each block has thickness 1 in the $x$ direction (i.e. pertains to a single parameter), its length in the $y$ direction is seq_length. In total we sample as many such blocks as the batch size is big. By construction those blocks are never the same throughout a training epoch but may intersect each other!

Library Functions

GeometricMachineLearning.DataLoader — Type

DataLoader(data)

Make an instance based on a data set.

This is designed to make training convenient.

Fields of DataLoader

The fields of the DataLoader struct are the following:

input: The input data with axes (i) system dimension, (ii) number of time steps and (iii) number of parameters.

output: The tensor that contains the output (supervised learning) - this may be of type Nothing if the constructor is only called with one tensor (unsupervised learning).

input_dim: The dimension of the system, i.e. what is taken as input by a regular neural network.

input_time_steps: The length of the entire time series (length of the second axis).

n_params: The number of parameters that are present in the data set (length of third axis)

output_dim: The dimension of the output tensor (first axis). If output is of type Nothing, then this is also of type Nothing.

output_time_steps: The size of the second axis of the output tensor. If output is of type Nothing, then this is also of type Nothing.

Implementation

Even though DataLoader can be called with inputs of various forms, internally it always stores tensors with three axes.

using GeometricMachineLearning

data = [1 2 3; 4 5 6]
dl = DataLoader(data)
dl.input

# output

[ Info: You have provided a matrix as input. The axes will be interpreted as (i) system dimension and (ii) number of parameters.
2×1×3 Array{Int64, 3}:
[:, :, 1] =
 1
 4

[:, :, 2] =
 2
 5

[:, :, 3] =
 3
 6

GeometricMachineLearning.Batch — Type

Batch

Batch is a struct whose functor acts on an instance of DataLoader to produce a sequence of training samples for training for one epoch.

See Batch(::Int) and Batch(::Int, ::Int, ::Int) for the different constructors.

The functor

An instance of Batch can be called on an instance of DataLoader to produce a sequence of samples that contain all the input data, i.e. for training for one epoch.

The output of applying batch:Batch to dl::DataLoader is a tuple of vectors of integers. Each of these vectors contains two integers: the first is the time index and the second one is the parameter index.

Examples

Consider the following example for drawing batches of size 2 for an instance of DataLoader constructed with a vector:

using GeometricMachineLearning
import Random

rng = Random.TaskLocalRNG()
Random.seed!(rng, 123)

dl = DataLoader(rand(rng, 5))
batch = Batch(2)

batch(dl)

# output

[ Info: You have provided a matrix as input. The axes will be interpreted as (i) system dimension and (ii) number of parameters.
([(1, 5), (1, 3)], [(1, 4), (1, 1)], [(1, 2)])

Here the first index is always 1 (the time dimension). We get a total number of 3 batches. The last batch is only of size 1 because we sample without replacement. Also see the docstring for DataLoader(::AbstractVector).

GeometricMachineLearning.Batch — Method

Batch(batch_size)

Make an instance of Batch for a specific batch size.

This is, among others, used to train neural networks of NeuralNetworkIntegrator type (as opposed to TransformerIntegrator).

GeometricMachineLearning.Batch — Method

Batch(batch_size, seq_length)

Make an instance of Batch for a specific batch size and a sequence length.

This is used to train neural networks of TransformerIntegrator type.

Optionally the prediction window can also be specified by calling:

using GeometricMachineLearning

batch_size = 2
seq_length = 3
prediction_window = 2

Batch(batch_size, seq_length, prediction_window)

# output

Batch{:Transformer}(2, 3, 2)

Note that here the batch is of type :Transformer.

GeometricMachineLearning.number_of_batches — Function

number_of_batches(dl, batch)

Compute the number of batches.

Here the distinction is between data that are time-series like and data that are autoencoder like.

Examples

using GeometricMachineLearning
using GeometricMachineLearning: number_of_batches
import Random

Random.seed!(123)

dat = [1, 2, 3, 4, 5]
dl₁ = DataLoader(dat; autoencoder = false, suppress_info = true) # time series-like
dl₂ = DataLoader(dat; autoencoder = true, suppress_info = true) # autoencoder-like
batch = Batch(3)

nob₁ = number_of_batches(dl₁, batch)
nob₂ = number_of_batches(dl₂, batch)
println(stdout, "Number of batches of dl₁: ", nob₁)
println(stdout, "Number of batches of dl₂: ", nob₂)
println(stdout, batch(dl₁), "\n", batch(dl₂))

# output

Number of batches of dl₁: 2
Number of batches of dl₂: 2
([(1, 1), (4, 1), (2, 1)], [(3, 1)])
([(1, 3), (1, 2), (1, 4)], [(1, 1), (1, 5)])

Here we see that in the autoencoder case that last minibatch has an additional element.

GeometricMachineLearning.convert_input_and_batch_indices_to_array — Function

convert_input_and_batch_indices_to_array(dl, batch, batch_indices)

Assign batch data based on (i) input and (ii) batch indices.

Examples

using GeometricMachineLearning

dl = DataLoader([0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9]; suppress_info = true)
batch = Batch(3)
batch_indices = [(1, 1), (1, 3), (1, 5)]

GeometricMachineLearning.convert_input_and_batch_indices_to_array(dl, batch, batch_indices)

# output

1×1×3 Array{Float64, 3}:
[:, :, 1] =
 0.1

[:, :, 2] =
 0.3

[:, :, 3] =
 0.5

1We first demonstrate how to sample data on the example of a matrix. The case of sampling from a tensor is slightly more complicated and is explained below.