GeometricMachineLearning Library Functions

GeometricMachineLearning.AbstractCache — Type

AbstractCache has subtypes: AdamCache MomentumCache GradientCache

All of them can be initialized with providing an array (also supporting manifold types).

GeometricMachineLearning.AbstractRetraction — Type

AbstractRetraction is a type that comprises all retraction methods for manifolds. For every manifold layer one has to specify a retraction method that takes the layer and elements of the (global) tangent space.

GeometricMachineLearning.AdamOptimizer — Type

Defines the Adam Optimizer. Algorithm and suggested defaults are taken from (Goodfellow et al., 2016, page 301), except for δ, because single precision is used!

GeometricMachineLearning.Attention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

GeometricMachineLearning.BlockIdentityLowerMatrix — Type

A BlockIdentityLowerMatrix is a matrix with blocks | 1 0 | | S 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.BlockIdentityUpperMatrix — Type

A BlockIdentityUpperMatrix is a matrix with blocks | 1 S | | 0 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.Classification — Type

Classification Layer that takes a matrix as an input and returns a vector that is used for MNIST classification.

TODO: Implement picking the last vector.

GeometricMachineLearning.DataLoader — Type

Data Loader is a struct that creates an instance based on a tensor (or different input format) and is designed to make training convenient.

Implemented:

If the data loader is called with a single tensor, a batchsize and an outputsize, then the batch is drawn randomly in the relevant range and the output is assigned accordingly.

The fields of the struct are the following:

   - data: The input data with axes (i) system dimension, (ii) number of parameters and (iii) number of time steps.

   - batch: A tensor in which the current batch is stored.

   - target_tensor: A tensor in which the current target is stored.

   - output: The tensor from which the output is drawn - this may be of type Nothing if the constructor is only called with one tensor.

   - sys_dim: The ``dimension'' of the system, i.e. what is taken as input by a regular neural network.

   - seq_length: The length batches should have.

   - batch_size:

   - output_size: The size of the second axis of the output tensor (prediction_window, output_size=1 in most cases)

   - n_params: The number of parameters that are present in the data set (length of second axis)

   - n_time_steps: Number of time steps (length of third axis)

For drawing the batch, the sampling is done over nparams and ntimesteps (here seqlength and output_size are also taken into account).

TODO: Implement DataLoader that works well with GeometricEnsembles etc.

GeometricMachineLearning.GradientOptimizer — Type

Define the Gradient optimizer, i.e. W ← W - η*∇f(W) Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.GrassmannLayer — Type

Defines a layer that performs simple multiplication with an element of the Grassmann manifold.

GeometricMachineLearning.GrassmannManifold — Type

maybe consider dividing the output in the check functions by n! TODO: Implement sampling procedures!!

GeometricMachineLearning.LayerWithManifold — Type

Additional types to make handling manifolds more readable.

GeometricMachineLearning.ManifoldLayer — Type

This defines a manifold layer that only has one matrix-valued manifold $A$ associated with it does $x\mapsto{}Ax$.

GeometricMachineLearning.MomentumOptimizer — Type

Define the Momentum optimizer, i.e. V ← αV - ∇f(W) W ← W + ηV Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.MultiHeadAttention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

GeometricMachineLearning.Optimizer — Type

Optimizer struct that stores the 'method' (i.e. Adam with corresponding hyperparameters), the cache and the optimization step.

It takes as input an optimization method and the parameters of a network.

GeometricMachineLearning.PSDLayer — Type

This is a PSD-like layer used for symplectic autoencoders. One layer has the following shape:

|Φ 0|

A = |0 Φ|, where Φ is an element of the regular Stiefel manifold.

GeometricMachineLearning.StiefelLayer — Type

Defines a layer that performs simple multiplication with an element of the Stiefel manifold.

GeometricMachineLearning.SymmetricMatrix — Type

A SymmetricMatrix is a matrix | a S | | S b |

The first index is the row index, the second one the column index.

If the constructor is called with a matrix as input it returns a symmetric matrix via the projection A ↦ .5*(A + Aᵀ). This is a projection defined via the canonical metric (A,B) ↦ tr(AᵀB).

TODO: Overload Adjoint operation for SymmetricMatrix!! (Aᵀ = A)

GeometricMachineLearning.SympNetLayer — Type

Implements the various layers from the SympNet paper: (https://www.sciencedirect.com/science/article/abs/pii/S0893608020303063). Its components are of the form:

\[\begin{pmatrix} I & \nabla{}V \\ 0 & I \end{pmatrix},\]

with $V(p) = \sum_ia_i\Sigma(\sum_jk_{ij}p_j+b_i)$, where $\Sigma$ is the antiderivative of the activation function $\sigma$ (one-layer neural network). Such layers are by construction symplectic.

For the linear layer, the activation and the bias are left out, and for the activation layer K and b are left out!

GeometricMachineLearning.TrainingData — Type

TrainingData stores:

 - problem 

 - shape 

 - get 

 - symbols 

 - dim 

 - noisemaker

GeometricMachineLearning.Transformer — Method

The architecture for a "transformer encoder" is essentially taken from arXiv:2010.11929, but with the difference that 𝐧𝐨 layer normalization is employed. This is because we still need to find a generalization of layer normalization to manifolds.

GeometricMachineLearning.assign_batch_kernel! — Method

Takes as input a $batch tensor'' (to which the data are assigned), the whole data tensor and two vectors$params'' and ``time_steps'' that include the specific parameters and time steps we want to assign.

Note that this assigns sequential data! For e.g. being processed by a transformer.

GeometricMachineLearning.assign_output_estimate — Method

Closely related to the transformer. It takes the last prediction_window columns of the output and uses is for the final prediction.

GeometricMachineLearning.assign_output_kernel! — Method

This should be used together with assignbatchkernel!. It assigns the corresponding output (i.e. target).

GeometricMachineLearning.augment_zeros_kernel! — Method

Used for differentiating assignoutputestimate (this appears in the loss).

GeometricMachineLearning.draw_batch! — Method

This function draws random time steps and parameters and based on these assign the batch and the output.

For ODE DataLoader: (i) batch input tensor to be written on (ii) batch output tensor to be written on (iii) data tensor

For MNIST DataLoader: (i) batch input tensor to be written on (ii) batch output tensor to be written on (iii) data tensor (iv) target tensor

GeometricMachineLearning.draw_batch! — Method

This assigns the batch if the data are in form of a matrix.

GeometricMachineLearning.init_optimizer_cache — Method

Wrapper for the functions setupadamcache, setupmomentumcache, setupgradientcache. These appear outside of optimizer_caches.jl because the OptimizerMethods first have to be defined.

GeometricMachineLearning.patch_index — Method

Based on coordinates i,j this returns the batch index (for MNIST data set for now).

GeometricMachineLearning.train! — Function

train!(...)

Perform a training of a neural networks on data using given method a training Method

Different ways of use:

train!(neuralnetwork, data, optimizer = GradientOptimizer(1e-2), training_method; nruns = 1000, batch_size = default(data, type), showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data : the data (see TrainingData)

optimizer = GradientOptimizer: the optimization method (see Optimizer)

training_method : specify the loss function used

nruns : number of iteration through the process with default value

batch_size : size of batch of data used for each step

GeometricMachineLearning.train! — Method

train!(neuralnetwork, data, optimizer, training_method; nruns = 1000, batch_size, showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data::AbstractTrainingData : the data

``

GeometricMachineLearning.within_patch_index — Method

Based on coordinates i,j this returns the index within the batch