GeometricMachineLearning Library Functions

GeometricMachineLearning.AbstractCache — Type

AbstractCache has subtypes: AdamCache MomentumCache GradientCache

All of them can be initialized with providing an array (also supporting manifold types).

GeometricMachineLearning.AbstractRetraction — Type

AbstractRetraction is a type that comprises all retraction methods for manifolds. For every manifold layer one has to specify a retraction method that takes the layer and elements of the (global) tangent space.

GeometricMachineLearning.AdamOptimizer — Type

Defines the Adam Optimizer. Algorithm and suggested defaults are taken from (Goodfellow et al., 2016, page 301), except for δ, because single precision is used!

GeometricMachineLearning.Attention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

GeometricMachineLearning.Batch — Type

Batch is a struct with an associated functor that acts on DataLoader.

The functor returns indices that are then used in the optimization step (always for an entire epoch).

GeometricMachineLearning.BlockIdentityLowerMatrix — Type

A BlockIdentityLowerMatrix is a matrix with blocks | 1 0 | | S 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.BlockIdentityUpperMatrix — Type

A BlockIdentityUpperMatrix is a matrix with blocks | 1 S | | 0 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.Classification — Type

Classification Layer that takes a matrix as an input and returns a vector that is used for MNIST classification.

It has the following arguments:

M: input dimension

N: output dimension

activation: the activation function

And the following optional argument:

average: If this is set to true, then the output is computed as $\frac{1}{N}\sum_{i=1}^N[input]_{\bullet{}i}$. If set to false (the default) it picks the last column of the input.

GeometricMachineLearning.DataLoader — Type

Data Loader is a struct that creates an instance based on a tensor (or different input format) and is designed to make training convenient.

Implemented:

If the data loader is called with a single tensor, a batchsize and an outputsize, then the batch is drawn randomly in the relevant range and the output is assigned accordingly.

The fields of the struct are the following:

data: The input data with axes (i) system dimension, (ii) number of parameters and (iii) number of time steps.

output: The tensor that contains the output (supervised learning) - this may be of type Nothing if the constructor is only called with one tensor (unsupervised learning).

input_dim: The ``dimension'' of the system, i.e. what is taken as input by a regular neural network.

input_time_steps: The length of the entire time series of the data

n_params: The number of parameters that are present in the data set (length of second axis)

output_dim: The dimension of the output tensor (first axis).

output_time_steps: The size of the second axis of the output tensor (also called predictionwindow, `outputsize=1` in most cases)

For drawing the batch, the sampling is done over nparams and ntimesteps (here seqlength and output_size are also taken into account).

If for the output we have a tensor whose second axis has length 1, we still store it as a tensor and not a matrix. This is because it is not necessarily of length 1.

TODO: Implement DataLoader that works well with GeometricEnsembles etc.

GeometricMachineLearning.DataLoader — Method

The DataLoader is called with a single tensor (snapshot tensor)

GeometricMachineLearning.DataLoader — Method

The DataLoader is called with a single matrix (snapshot matrix)

GeometricMachineLearning.DataLoader — Method

The DataLoader is called with a tensor and a vector. For the moment this is always interpreted to be the MNIST data set.

GeometricMachineLearning.GradientOptimizer — Type

Define the Gradient optimizer, i.e. W ← W - η*∇f(W) Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.GrassmannLayer — Type

Defines a layer that performs simple multiplication with an element of the Grassmann manifold.

GeometricMachineLearning.GrassmannManifold — Type

maybe consider dividing the output in the check functions by n! TODO: Implement sampling procedures!!

GeometricMachineLearning.LayerWithManifold — Type

Additional types to make handling manifolds more readable.

GeometricMachineLearning.ManifoldLayer — Type

This defines a manifold layer that only has one matrix-valued manifold $A$ associated with it does $x\mapsto{}Ax$.

GeometricMachineLearning.MomentumOptimizer — Type

Define the Momentum optimizer, i.e. V ← αV - ∇f(W) W ← W + ηV Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.MultiHeadAttention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

GeometricMachineLearning.Optimizer — Type

Optimizer struct that stores the 'method' (i.e. Adam with corresponding hyperparameters), the cache and the optimization step.

It takes as input an optimization method and the parameters of a network.

GeometricMachineLearning.PSDLayer — Type

This is a PSD-like layer used for symplectic autoencoders. One layer has the following shape:

\[A = \begin{bmatrix} \Phi & \mathbb{O} \\ \mathbb{O} & \Phi \end{bmatrix},\]

where $\Phi$ is an element of the Stiefel manifold $St(n, N)$.

The constructor of PSDLayer is called by PSDLayer(M, N; retraction=retraction):

M is the input dimension.

N is the output dimension.

retraction is an instance of a struct with supertype AbstractRetraction. The only options at the moment are Geodesic() and Cayley().

GeometricMachineLearning.ReducedSystem — Type

ReducedSystem computes the reconstructed dynamics in the full system based on the reduced one. Optionally it can be compared to the FOM solution.

It can be called using the following constructor: ReducedSystem(N, n, encoder, decoder, fullvectorfield, reducedvectorfield, params, tspan, tstep, ics, projection_error) where

encoder: a function $\mathbb{R}^{2N}\mapsto{}\mathbb{R}^{2n}$

decoder: a (differentiable) function $\mathbb{R}^{2n}\mapsto\mathbb{R}^{2N}$

fullvectorfield: a (differentiable) mapping defined the same way as in GeometricIntegrators

reducedvectorfield: a (differentiable) mapping defined the same way as in GeometricIntegrators

params: a NamedTuple that parametrizes the vector fields (the same for fullvectorfield and reducedvectorfield)

tspan: a tuple (t₀, tₗ) that specifies start and end point of the time interval over which integration is performed.

tstep: the time step

ics: the initial condition for the big system.

projection_error: the error $||M - \mathcal{R}\circ\mathcal{P}(M)||$ where $M$ is the snapshot matrix; $\mathcal{P}$ and $\mathcal{R}$ are the reduction and reconstruction respectively.

GeometricMachineLearning.StiefelLayer — Type

Defines a layer that performs simple multiplication with an element of the Stiefel manifold.

GeometricMachineLearning.SymmetricMatrix — Type

A SymmetricMatrix is a matrix | a S | | S b |

The first index is the row index, the second one the column index.

If the constructor is called with a matrix as input it returns a symmetric matrix via the projection A ↦ .5*(A + Aᵀ). This is a projection defined via the canonical metric (A,B) ↦ tr(AᵀB).

TODO: Overload Adjoint operation for SymmetricMatrix!! (Aᵀ = A)

GeometricMachineLearning.SympNetLayer — Type

Implements the various layers from the SympNet paper: (https://www.sciencedirect.com/science/article/abs/pii/S0893608020303063). Its components are of the form:

\[\begin{pmatrix} I & \nabla{}V \\ 0 & I \end{pmatrix},\]

with $V(p) = \sum_ia_i\Sigma(\sum_jk_{ij}p_j+b_i)$, where $\Sigma$ is the antiderivative of the activation function $\sigma$ (one-layer neural network). Such layers are by construction symplectic.

For the linear layer, the activation and the bias are left out, and for the activation layer K and b are left out!

GeometricMachineLearning.SystemType — Type

Can specify a special type of the system, to be used with ReducedSystem. For now the only option is Symplectic (and NoStructure).

GeometricMachineLearning.TrainingData — Type

TrainingData stores:

 - problem 

 - shape 

 - get 

 - symbols 

 - dim 

 - noisemaker

GeometricMachineLearning.Transformer — Method

The architecture for a "transformer encoder" is essentially taken from arXiv:2010.11929, but with the difference that no layer normalization is employed. This is because we still need to find a generalization of layer normalization to manifolds.

The transformer is called with the following inputs:

dim: the dimension of the transformer

n_heads: the number of heads

L: the number of transformer blocks

In addition we have the following optional arguments:

activation: the activation function used for the ResNet (tanh by default)

Stiefel::Bool: if the matrices $P^V$, $P^Q$ and $P^K$ should live on a manifold (false by default)

retraction: which retraction should be used (Geodesic() by default)

add_connection::Bool: if the input should by added to the ouput after the MultiHeadAttention layer is used (true by default)

use_bias::Bool: If the ResNet should use a bias (true by default)

GeometricMachineLearning.accuracy — Method

Computes the accuracy (as opposed to the loss) of a neural network classifier.

It takes as input:

model::Chain:

ps: parameters of the network

dl::DataLoader

GeometricMachineLearning.assign_batch_kernel! — Method

Takes as input a $batch tensor'' (to which the data are assigned), the whole data tensor and two vectors$params'' and ``time_steps'' that include the specific parameters and time steps we want to assign.

Note that this assigns sequential data! For e.g. being processed by a transformer.

GeometricMachineLearning.assign_output_estimate — Method

The function assign_output_estimate is closely related to the transformer. It takes the last prediction_window columns of the output and uses is for the final prediction. i.e.

\[\mathbb{R}^{N\times\mathtt{pw}}\to\mathbb{R}^{N\times\mathtt{pw}}, \begin{bmatrix} z^{(1)}_1 & \cdots z^{(T)}_1 \\ \cdots & \cdots \\ z^{(T - \mathtt{pw})}_n & \cdots & z^{(T})_n\end{bmatrix} \mapsto \begin{bmatrix} z^{(1)}_1 & \cdots z^{(T)}_1 \\ \cdots & \cdots \\ z^{(T - \mathtt{pw})}_n & \cdots & z^{(T})_n\end{bmatrix} \]

GeometricMachineLearning.assign_output_kernel! — Method

This should be used together with assignbatchkernel!. It assigns the corresponding output (i.e. target).

GeometricMachineLearning.augment_zeros_kernel! — Method

Used for differentiating assignoutputestimate (this appears in the loss).

GeometricMachineLearning.draw_batch! — Method

This function draws random time steps and parameters and based on these assign the batch and the output.

For ODE DataLoader: (i) batch input tensor to be written on (ii) batch output tensor to be written on (iii) data tensor

For MNIST DataLoader: (i) batch input tensor to be written on (ii) batch output tensor to be written on (iii) data tensor (iv) target tensor

GeometricMachineLearning.draw_batch! — Method

This assigns the batch if the data are in form of a matrix.

GeometricMachineLearning.init_optimizer_cache — Method

Wrapper for the functions setupadamcache, setupmomentumcache, setupgradientcache. These appear outside of optimizer_caches.jl because the OptimizerMethods first have to be defined.

GeometricMachineLearning.loss — Method

Computes the loss for a neural network and a data set. The computed loss is $||output - \mathcal{NN}(input)||_F/\mathtt{size(output, 2)}/\mathtt{size(output, 3)}$, where $||A||_F := \sqrt{\sum_{i_1,\ldots,i_k}||a_{i_1,\ldots,i_k}^2}$ is the Frobenius norm.

It takes as input:

model

ps

input

output

GeometricMachineLearning.loss — Method

Alternative call of the loss function. This takes as input:

model

ps

dl::DataLoader

GeometricMachineLearning.onehotbatch — Method

One-hot-batch encoding of a vector of integers: $input\in\{0,1,\ldots,9\}^\ell$. The output is a tensor of shape $10\times1\times\ell$.

\[0 \mapsto \begin{bmatrix} 1 & 0 & \ldots & 0 \end{matrix}.\]

In more abstract terms: $i \mapsto e_i$.

GeometricMachineLearning.optimization_step! — Method

Optimization for a single layer.

inputs:

o::Optimizer

d::Union{AbstractExplicitLayer, AbstractExplicitCell}

ps::NamedTuple: the parameters

C::NamedTuple: NamedTuple of the caches

dx::NamedTuple: NamedTuple of the derivatives (output of AD routine)

ps, C and dx must have the same keys.

GeometricMachineLearning.optimize_for_one_epoch! — Method

Optimize for an entire epoch. For this you have to supply:

an instance of the optimizer.

the neural network model

the parameters of the model

the data (in form of DataLoader)

in instance of Batch that contains batch_size (and optionally seq_length)

With the optional argument:

the loss, which takes the model, the parameters ps and an instance of DataLoader as input.

The output of optimize_for_one_epoch! is the average loss over all batches of the epoch:

\[output = \frac{1}{mathtt{steps\_per\_epoch}}\sum_{t=1}^mathtt{steps\_per\_epoch}loss(\theta^{(t-1)}).\]

This is done because any reverse differentiation routine always has two outputs: a pullback and the value of the function it is differentiating. In the case of zygote: loss_value, pullback = Zygote.pullback(ps -> loss(ps), ps) (if the loss only depends on the parameters).

GeometricMachineLearning.patch_index — Method

Based on coordinates i,j this returns the batch index (for MNIST data set for now).

GeometricMachineLearning.reduced_vector_field_from_full_explicit_vector_field — Method

This function is needed if we obtain a GeometricIntegrators-like vector field from an explicit vector field $V:\mathbb{R}^{2N}\to\mathbb{R}^{2N}$. We need this function because buildreducedvector_field is not working in conjunction with implicit integrators.

GeometricMachineLearning.split_and_flatten — Method

split_and_flatten takes a tensor as input and produces another one as output (essentially rearranges the input data in an intricate way) so that it can easily be processed with a transformer.

The optional arguments are:

patch_length: by default this is 7.

number_of_patches: by default this is 16.

GeometricMachineLearning.train! — Function

train!(...)

Perform a training of a neural networks on data using given method a training Method

Different ways of use:

train!(neuralnetwork, data, optimizer = GradientOptimizer(1e-2), training_method; nruns = 1000, batch_size = default(data, type), showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data : the data (see TrainingData)

optimizer = GradientOptimizer: the optimization method (see Optimizer)

training_method : specify the loss function used

nruns : number of iteration through the process with default value

batch_size : size of batch of data used for each step

GeometricMachineLearning.train! — Method

train!(neuralnetwork, data, optimizer, training_method; nruns = 1000, batch_size, showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data::AbstractTrainingData : the data

``

GeometricMachineLearning.within_patch_index — Method

Based on coordinates i,j this returns the index within the batch