GeometricMachineLearning Library Functions

GeometricMachineLearning.default_retr — Constant

The architecture for a "transformer encoder" is essentially taken from arXiv:2010.11929, but with the difference that 𝐧𝐨 layer normalization is employed. This is because we still need to find a generalization of layer normalization to manifolds.

GeometricMachineLearning.BlockIdentityLowerMatrix — Type

A BlockIdentityLowerMatrix is a matrix with blocks | 1 0 | | S 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.BlockIdentityUpperMatrix — Type

A BlockIdentityUpperMatrix is a matrix with blocks | 1 S | | 0 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.GradientOptimizer — Type

Define the Gradient optimizer, i.e. W ← W - η*∇f(W) Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.GrassmannManifold — Type

maybe consider dividing the output in the check functions by n! TODO: Implement sampling procedures!!

GeometricMachineLearning.MomentumOptimizer — Type

Define the Momentum optimizer, i.e. V ← αV - ∇f(W) W ← W + ηV Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.NeuralNetworkLayer — Type

A NeuralNetworkLayer is a map from $\mathbb{R}^{N} \rightarrow \mathbb{R}^{M}$.

GeometricMachineLearning.SymplecticHouseholderDecom — Type

this algorithm is taken (and adjusted) from https://doi.org/10.1016/j.laa.2008.02.029

GeometricMachineLearning.chain — Function

The function chain returns a chain of a specific neural network architecture and a specific backend and is called by:

chain(::AbstractArchitecture, ::AbstractBackend)

GeometricMachineLearning.train! — Function

train!(...)

Perform a training of a neural networks on data using given method a training integrator

Different ways of use:

train!(neuralnetwork, data, optimizer = GradientOptimizer(1e-2), training_method; nruns = 1000, batch_size = default(data, type), showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data : the data (see TrainingData)

optimizer = GradientOptimizer: the optimization method (see Optimizer)

training_method : specify the loss function used

nruns : number of iteration through the process with default value

batch_size : size of batch of data used for each step

GeometricMachineLearning.train! — Method

train!(neuralnetwork, data, optimizer, training_method; nruns = 1000, batch_size, showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data::AbstractTrainingData : the data

``