GeometricMachineLearning Library Functions

GeometricMachineLearning.AbstractCache — Type

AbstractCache has subtypes: AdamCache MomentumCache GradientCache

All of them can be initialized with providing an array (also supporting manifold types).

GeometricMachineLearning.AbstractRetraction — Type

AbstractRetraction is a type that comprises all retraction methods for manifolds. For every manifold layer one has to specify a retraction method that takes the layer and elements of the (global) tangent space.

GeometricMachineLearning.AdamOptimizer — Type

Defines the Adam Optimizer. Algorithm and suggested defaults are taken from (Goodfellow et al., 2016, page 301), except for δ, because single precision is used!

GeometricMachineLearning.BlockIdentityLowerMatrix — Type

A BlockIdentityLowerMatrix is a matrix with blocks | 1 0 | | S 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.BlockIdentityUpperMatrix — Type

A BlockIdentityUpperMatrix is a matrix with blocks | 1 S | | 0 1 | Currently, it only implements a custom mul! method, exploiting this structure.

GeometricMachineLearning.Gradient — Type

The gradient layer from the SympNet paper (https://www.sciencedirect.com/science/article/abs/pii/S0893608020303063). Its components are of the form: $ \begin{pmatrix} I & \nabla{}V \ 0 & I \end{pmatrix}, $ with $V(p) = \sum_ia_i\Sigma(\sum_jk_{ij}p_j+b_i)$, where $\Sigma$ is the antiderivative of the activation function $\sigma$. Such layers are by construction symplectic.

GeometricMachineLearning.GradientOptimizer — Type

Define the Gradient optimizer, i.e. W ← W - η*∇f(W) Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.GrassmannLayer — Type

Defines a layer that performs simple multiplication with an element of the Grassmann manifold.

GeometricMachineLearning.GrassmannManifold — Type

maybe consider dividing the output in the check functions by n! TODO: Implement sampling procedures!!

GeometricMachineLearning.ManifoldLayer — Type

This defines a manifold layer that only has one matrix-valued manifold $A$ associated with it does $x\mapsto{}Ax$.

GeometricMachineLearning.MomentumOptimizer — Type

Define the Momentum optimizer, i.e. V ← αV - ∇f(W) W ← W + ηV Or the riemannian manifold equivalent, if applicable.

GeometricMachineLearning.MultiHeadAttention — Type

MultiHeadAttention (MHA) serves as a preprocessing step in the transformer. It reweights the input vectors bases on correlations within those data.

GeometricMachineLearning.Optimizer — Type

Optimizer struct that stores the 'method' (i.e. Adam with corresponding hyperparameters), the cache and the optimization step.

It takes as input an optimization method and the parameters of a network.

GeometricMachineLearning.StiefelLayer — Type

Defines a layer that performs simple multiplication with an element of the Stiefel manifold.

GeometricMachineLearning.SymplecticHouseholderDecom — Type

this algorithm is taken (and adjusted) from https://doi.org/10.1016/j.laa.2008.02.029

GeometricMachineLearning.Transformer — Method

The architecture for a "transformer encoder" is essentially taken from arXiv:2010.11929, but with the difference that 𝐧𝐨 layer normalization is employed. This is because we still need to find a generalization of layer normalization to manifolds.

GeometricMachineLearning.init_optimizer_cache — Method

Wrapper for the functions setupadamcache, setupmomentumcache, setupgradientcache. These appear outside of optimizer_caches.jl because the OptimizerMethods first have to be defined.

GeometricMachineLearning.train! — Function

train!(...)

Perform a training of a neural networks on data using given method a training Method

Different ways of use:

train!(neuralnetwork, data, optimizer = GradientOptimizer(1e-2), training_method; nruns = 1000, batch_size = default(data, type), showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data : the data (see TrainingData)

optimizer = GradientOptimizer: the optimization method (see Optimizer)

training_method : specify the loss function used

nruns : number of iteration through the process with default value

batch_size : size of batch of data used for each step

GeometricMachineLearning.train! — Method

train!(neuralnetwork, data, optimizer, training_method; nruns = 1000, batch_size, showprogress = false )

Arguments

neuralnetwork::LuxNeuralNetwork : the neural net work using LuxBackend

data::AbstractTrainingData : the data

``